% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_graph.R
\name{addColourAttributeFr}
\alias{addColourAttributeFr}
\title{Add colour attribute to nodes in graphml file}
\usage{
addColourAttributeFr(simp, graphfilename, frattributename = "FR",
  colourattributename = "ncolor")
}
\arguments{
\item{simp}{}

\item{graphfilename}{}

\item{frattributename}{}

\item{colourattributename}{}
}
\value{
new graph file
}
\description{
Add colour attribute to nodes in graphml file
}
\author{
Sascha Holzhauer
}
