% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_configuration.R
\name{adjust_changecolumnnames}
\alias{adjust_changecolumnnames}
\title{Change column names of CSV files, e.g. capital files}
\usage{
adjust_changecolumnnames(sip, indir = simp$dirs$param$getparamdir(simp,
  "capitals"), outdir = indir, colname_old, colname_new,
  filepattern = "*.csv")
}
\arguments{
\item{sip}{}

\item{indir}{}

\item{outdir}{}

\item{colname_old}{}

\item{colname_new}{}

\item{filepattern}{}
}
\value{
adjusted files
}
\description{
Change column names of CSV files, e.g. capital files
}
\author{
Sascha Holzhauer
}
