% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_sa.R
\name{analyse_statistics_sa_local}
\alias{analyse_statistics_sa_local}
\title{Spatial autocorrelation score for given raster}
\usage{
analyse_statistics_sa_local(raster, data = c(1), type = c("Moran", "Geary",
  "MoranLocal", "GearyLocal"))
}
\arguments{
\item{raster}{}

\item{data}{}

\item{type}{}
}
\value{
sa score
}
\description{
Spatial autocorrelation score for given raster
}
\author{
Sascha Holzhauer
}
