% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_2raster}
\alias{convert_2raster}
\title{Convert data.frame with coordinates into raster object}
\usage{
convert_2raster(simp, data, targetCRS = "+init=EPSG:32632", layers = c(1))
}
\arguments{
\item{simp}{SIMulation Properties}

\item{data}{list of data containing columns x and y and data colum(s)}

\item{layers}{vector or list of layer indices or names. Length may not exceed number of data columns.}
}
\value{
list (e.g. regions) of lists (e.g. capitals) of raster obejcts
}
\description{
Convert data.frame with coordinates into raster object
}
\author{
Sascha Holzhauer
}

