% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_aggregate.R
\name{convert_aggregate_demand}
\alias{convert_aggregate_demand}
\title{Aggregate and store demand}
\usage{
convert_aggregate_demand(simp, checkexists = TRUE, forceparam = FALSE,
  demanddataname = "csv_aggregateServiceDemand",
  sourcedataname = "dataAggregateSupplyDemand")
}
\arguments{
\item{simp}{considered elements: \itemize{
\item{\code{simp$mdata$conversion$services}}
\item{\code{\link{input_tools_load}}}
\item{\code{\link{input_csv_param_demand}}}}}

\item{forceparam}{don't attempt to use stored rData.}

\item{demanddataname}{}
}
\value{
-
}
\description{
Tries to load demand from stored rData (usually from separate supply and demand CSV outputs).
If rData not present at \code{sourcedataname} it uses param data.
}
\author{
Sascha Holzhauer
}
