% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_aggregate_meltsupplydemand}
\alias{convert_aggregate_meltsupplydemand}
\title{Melts demand and supply data and splits service columns into type and service}
\usage{
convert_aggregate_meltsupplydemand(simp, data)
}
\arguments{
\item{simp}{}

\item{data}{}
}
\value{
melted data.frame
}
\description{
Melts demand and supply data and splits service columns into type and service
}
\author{
Sascha Holzhauer
}

