% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_aggregate_supply}
\alias{convert_aggregate_supply}
\title{Aggregate and store supply}
\usage{
convert_aggregate_supply(simp, dataname = "csv_cell_aggregated",
  checkexists = TRUE, supplydataname = "csv_aggregated_supply")
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{checkexists}{}

\item{demanddataname}{}
}
\value{
-
}
\description{
Aggregate and store supply
}
\author{
Sascha Holzhauer
}

