% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_aggregate_takeovers}
\alias{convert_aggregate_takeovers}
\title{Extracts numbers of take overs for every pair of AFT for every tick from stored cell csv data}
\usage{
convert_aggregate_takeovers(simp,
  landusedataname = "csv_LandUseIndex_rbinded", grouping = c("Scenario",
  "Runid", "Region"))
}
\arguments{
\item{simp}{}

\item{landusedataname}{}

\item{grouping}{adjusts which columns are considered for aggregatation}
}
\value{
data.frame with previous AFT as row and resulting AFT as column
}
\description{
Scenario, Runid and Region are preserved.
}
\author{
Sascha Holzhauer
}

