% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_raster_aggregate.R
\name{convert_raster_flatlist}
\alias{convert_raster_flatlist}
\title{Aggregate raster data with the given aggregate function}
\usage{
convert_raster_flatlist(inforasterdata, aggregatefunction = "sum")
}
\arguments{
\item{inforasterdata}{}

\item{aggregatefunction}{}
}
\value{
data.frame
}
\description{
Aggregate raster data with the given aggregate function
}
\author{
Sascha Holzhauer
}

