% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_raster_getAftNumbers}
\alias{convert_raster_getAftNumbers}
\title{Compute AFt numbers}
\usage{
convert_raster_getAftNumbers(aft_raster)
}
\arguments{
\item{aft_raster}{}
}
\value{
data.farme with numbers and info
}
\description{
Compute AFt numbers
}
\author{
Sascha Holzhauer
}

