% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_aggregate.R
\name{convert_supplydemand_percentage}
\alias{convert_supplydemand_percentage}
\title{Calculates the percental demand supply gap for all ticks}
\usage{
convert_supplydemand_percentage(simp,
  datanamedemand = "csv_aggregated_demand",
  datanamesupply = "csv_aggregated_supply", includesum = TRUE,
  asvector = FALSE)
}
\arguments{
\item{simp}{}

\item{includesum}{if \code{TRUE} computes the sum across all services}

\item{dataname}{}
}
\value{
named (ticks) vector of percental demand supply gap
}
\description{
Calculates the percental demand supply gap for all ticks
}
\author{
Sascha Holzhauer
}

