% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_xmlfunction_parameter}
\alias{get_xmlfunction_parameter}
\title{Extact parameter from curce XML file}
\usage{
get_xmlfunction_parameter(data, rundata, parameter, default)
}
\arguments{
\item{data}{}

\item{rundata}{}

\item{parameter}{}

\item{default}{}
}
\value{
parameter value
}
\description{
Considers given default value in case the parameter is not defined in the XML file.
}
\author{
Sascha Holzhauer
}

