% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_configuration.R
\name{getCompetitionFunctionParams}
\alias{getCompetitionFunctionParams}
\title{Calculate optimal linear competition function parameters (according to document
"How to calibrate competition functions in CRAFTY?")}
\usage{
getCompetitionFunctionParams(simp, production = NULL, producingCells = NULL,
  paramAfactor = 0.75, paramBfactor = 1,
  datanameAFTcomposition = "csv_aggregateAFTComposition")
}
\arguments{
\item{simp}{simp$sim$id must specify the run to source production from}
}
\value{
data.frame of parameters
}
\description{
Calculate optimal linear competition function parameters (according to document
"How to calibrate competition functions in CRAFTY?")
}
\author{
Sascha Holzhauer
}

