% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_various.R
\name{hl_actions_fillDoNothing}
\alias{hl_actions_fillDoNothing}
\title{Fill given data with \code{doNothingAction} for missing entries of the combinations
Tick/Runid/Region/Agent}
\usage{
hl_actions_fillDoNothing(simp, data, doNothingAction = "DoNothing",
  score = 0)
}
\arguments{
\item{simp}{}

\item{data}{}

\item{doNothingAction}{}

\item{score}{the score value that is assigned to added DoNothing actions.}
}
\value{
manipulated data.frame
}
\description{
Fill given data with \code{doNothingAction} for missing entries of the combinations
Tick/Runid/Region/Agent
}
\author{
Sascha Holzhauer
}
