% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_aftcomposisition_file}
\alias{hl_aftcomposisition_file}
\title{Reads single CSV file with AFT composition and produces timeline figure}
\usage{
hl_aftcomposisition_file(simp, csvfilename, title = "AftComposition",
  figurefilename = "AftComposition")
}
\arguments{
\item{simp}{}

\item{csvfilename}{}

\item{title}{}

\item{figurefilename}{}
}
\value{
figure
}
\description{
Reads single CSV file with AFT composition and produces timeline figure
}
\author{
Sascha Holzhauer
}

