% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_aftmap_changes}
\alias{hl_aftmap_changes}
\title{Plot changes between ticks for a specific (combination of) land use(s)}
\usage{
hl_aftmap_changes(simp, cdata, ncol = 1, title = "AFT-Changes",
  ggplotaddon = NULL, regions = simp$sim$regions,
  addcountryshapes = FALSE, plotunchanged = TRUE)
}
\arguments{
\item{simp}{}

\item{cdata}{list of celldata to compare}

\item{ncol}{}

\item{ggplotaddon}{}
}
\value{
figure file / plot
}
\description{
Plot changes between ticks for a specific (combination of) land use(s)
}
\author{
Sascha Holzhauer
}

