% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_maps.R
\name{hl_aftmap_multi}
\alias{hl_aftmap_multi}
\title{AFT map for 2010 and 2040 with differences (placed horicontally) for multiple Runids}
\usage{
hl_aftmap_multi(simp, dataname = "csv_LandUseIndex_rbinded",
  firsttick = 2010, secondtick = 2040, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
map plot
}
\description{
AFT map for 2010 and 2040 with differences (placed horicontally) for multiple Runids
}
\author{
Sascha Holzhauer
}
