% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_afttakeoverfluctuations}
\alias{hl_afttakeoverfluctuations}
\title{AFT take over fluctations as timeline}
\usage{
hl_afttakeoverfluctuations(simp, dataname = "csv_cell_aggregated",
  starttick = simp$sim$starttick, tickinterval = 5,
  endtick = simp$sim$endtick, datanametakeovers = "csv_aggregateTakeOver",
  returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{starttick}{}

\item{tickinterval}{}

\item{endtick}{}

\item{datanametakeovers}{}

\item{returnplot}{if true the ggplot object is returned}

\item{runid}{}
}
\value{
timeline plot
}
\description{
AFT take over fluctations as timeline
}
\author{
Sascha Holzhauer
}
