% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_aggregate_aftcompetitiveness}
\alias{hl_aggregate_aftcompetitiveness}
\title{Yearly aggregated AFT competitiveness}
\usage{
hl_aggregate_aftcompetitiveness(simp,
  dataname = "csv_aggregateAFTCompetitiveness", includeunmanaged = FALSE,
  aggcelldataname = "dataAgg", returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{includeunmanaged}{if \code{TRUE} the number of unmanaged cells is plotted, too. It is derived from the
total number of cells obtained}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
timeline plot
}
\description{
Yearly aggregated AFT competitiveness
}
\author{
Sascha Holzhauer
}

