% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_aggregate_demand}
\alias{hl_aggregate_demand}
\title{Yearly aggregated demand}
\usage{
hl_aggregate_demand(simp, dataname = "csv_aggregated_demand",
  ggplotaddons = NULL, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
timeline plot
}
\description{
Yearly aggregated demand
}
\author{
Sascha Holzhauer
}
