% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_aggregate_demandsupply}
\alias{hl_aggregate_demandsupply}
\title{Yearly aggregated demand and supply}
\usage{
hl_aggregate_demandsupply(simp, dataname = "csv_aggregateServiceDemand")
}
\arguments{
\item{simp}{}

\item{dataname}{}
}
\value{
timeline plot
}
\description{
Yearly aggregated demand and supply
}
\author{
Sascha Holzhauer
}

