% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_aggregate_demandsupply_csv}
\alias{hl_aggregate_demandsupply_csv}
\title{Aggregate demand and supply for all regions per tick and store as CSV file.}
\usage{
hl_aggregate_demandsupply_csv(simp, dataname = "dataAggregateSupplyDemand")
}
\arguments{
\item{simp}{}

\item{dataname}{}
}
\value{
csv file
}
\description{
Aggregate demand and supply for all regions per tick and store as CSV file.
}
\author{
Sascha Holzhauer
}

