% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_aggregate_sa}
\alias{hl_aggregate_sa}
\title{Read stored landuse data, calculate and visualise spatial autocorrelation score}
\usage{
hl_aggregate_sa(simp, celldataname = "csv_LandUseIndex_rbinded",
  starttick = simp$sim$starttick, tickinterval = 10,
  endtick = simp$sim$endtick, linetypecol = "Region", type = "Moran",
  regions = simp$sim$regions, colourcol = NULL,
  titleprefix = paste("SpatialAutocorrelation (", type, ")", sep = ""),
  filenameprefix = paste("SpatialAutocorrelation", type, sep = "_"),
  returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{celldataname}{}

\item{starttick}{}

\item{tickinterval}{}

\item{endtick}{}

\item{linetypecol}{}

\item{type}{}

\item{regions}{}

\item{colourcol}{}

\item{titleprefix}{}

\item{filenameprefix}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
lines plot
}
\description{
NOTE: Needs the cell data to be stored as rData (see \link{input_csv_data}).
}
\author{
Sascha Holzhauer
}
