% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_cell2raster.R
\name{hl_cell2raster}
\alias{hl_cell2raster}
\title{Extracts cell coordinates and land use from stored cell data and generates raster file}
\usage{
hl_cell2raster(simp, tick, dataname = "csv_LandUseIndex_rbinded",
  landuseindexcolname = "LandUseIndex", naflag = -9)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{landuseindexcolname}{}

\item{naflag}{}
}
\value{
raster file
}
\description{
Extracts cell coordinates and land use from stored cell data and generates raster file
}
\author{
Sascha Holzhauer
}

