% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_comp_cell_aftcomposition}
\alias{hl_comp_cell_aftcomposition}
\title{Load, aggregate and visualise AFT composition data of several runs to compare}
\usage{
hl_comp_cell_aftcomposition(simp, simps, dataname = "csv_cell_aggregated")
}
\arguments{
\item{simp}{}

\item{simps}{list of simp that shall be compared. \code{simp$sim$shortid} is combined with Runid to distinguish data!}

\item{dataname}{name of aggregated data from cell data}
}
\value{
timeline plot
}
\description{
Load, aggregate and visualise AFT composition data of several runs to compare
}
\author{
Sascha Holzhauer
}

