% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_comp_aggregate.R
\name{hl_comp_demandsupplygap_agentparams}
\alias{hl_comp_demandsupplygap_agentparams}
\title{Supply and Demand gap depending on AFT param}
\usage{
hl_comp_demandsupplygap_agentparams(simp,
  simps = input_tools_buildsimplist(111:112),
  dataname = "csv_cell_aggregated", filename = paste("SupplyDemandGap_", if
  (!is.null(simp$sim$rundesc)) paste(simp$sim$rundesc, collapse = "-") else
  simp$sim$id, sep = ""), title = paste("Demand/Supply Gap", if
  (!is.null(paste(simp$sim$rundesc))) paste(simp$sim$rundesc, collapse = "/"),
  sep = " - "), agentparam = "givingUpProb", aft = simp$mdata$aftNames[2],
  ggplotparams = NULL, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{simps}{}

\item{dataname}{}

\item{filename}{}

\item{title}{}

\item{agentparam}{}

\item{aft}{}

\item{ggplotparams}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot
}
\description{
Plots the gap between demand and supply for various runs with
agent param as dependent variable.
}
\author{
Sascha Holzhauer
}

