% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_competitiveness}
\alias{hl_competitiveness}
\title{Load from csv data, aggregate, and plot AFT competitiveness}
\usage{
hl_competitiveness(simp, dataname = "csv_cell_aggregated")
}
\arguments{
\item{simp}{}

\item{dataname}{}
}
\value{
plot
}
\description{
Load from csv data, aggregate, and plot AFT competitiveness
}
\author{
Sascha Holzhauer
}

