% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_competitiveness_prealloc}
\alias{hl_competitiveness_prealloc}
\title{Visualise pre-allocation competitiveness per AFT}
\usage{
hl_competitiveness_prealloc(simp, dataname = "csv_preAlloc_rbinded",
  maxcompetitiveness = "100\%", numbins = 20, facet_ncol = NULL,
  filename = paste("PreAllocationCompetition", simp$sim$id, sep = "_"),
  title = NULL, ggplotaddons = NULL, setfigdims = TRUE,
  afts = simp$mdata$aftNames, checkexists = FALSE, skipemptybins = TRUE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{maxcompetitiveness}{either an absoulte value or a percentage given as string (e.g. "90%"). The latter
selects the lowest X percent.}

\item{numbins}{number of bins to bin data}

\item{facet_ncol}{if \code{NULL} ticks will be as rows, agents as columns. Otherwise, facets will
distinguished by ticks only (and facet_cols specifies the number of columns)}

\item{filename}{}

\item{title}{plot title}

\item{ggplotaddons}{list of ggplot2 elements}

\item{setfigdims}{if \code{TRUE} \code{simp$fig$height} and \code{simp$fig$width} are set appropriately}

\item{afts}{AFTs to display (as vector of names)}
}
\value{
facet histogram plot
}
\description{
Visualise pre-allocation competitiveness per AFT
}
\author{
Sascha Holzhauer
}

