% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_competitivenessPerRegion}
\alias{hl_competitivenessPerRegion}
\title{Load from csv data, aggregate, and plot AFT competitiveness per region}
\usage{
hl_competitivenessPerRegion(simp, dataname = "csv_cell_aggregated")
}
\arguments{
\item{simp}{}

\item{dataname}{}
}
\value{
ggplot2 plot
}
\description{
Load from csv data, aggregate, and plot AFT competitiveness per region
}
\author{
Sascha Holzhauer
}

