% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_compilerunparams}
\alias{hl_compilerunparams}
\title{Print LaTeX table including run parameters}
\usage{
hl_compilerunparams(simp, runidcolumnname = "run")
}
\arguments{
\item{simp}{}

\item{runidcolumnname}{}
}
\value{
xtable plot
}
\description{
Considers the \code{simp$sim$version} in column run.
}
\author{
Sascha Holzhauer
}

