% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_connectedness}
\alias{hl_connectedness}
\title{Visualise stored metric LandUseConnectivity as timelines}
\usage{
hl_connectedness(simp, dataname = "csv_aggregated_connectivity",
  datatype = "LandUseConnectivity", aftcolumns = simp$mdata$aftNames[-1],
  percent = NULL)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{datatype}{}

\item{datacolumns}{}
}
\value{
plot (and rData)
}
\description{
Visualise stored metric LandUseConnectivity as timelines
}
\author{
Sascha Holzhauer
}

