% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_demandsupply}
\alias{hl_demandsupply}
\title{Read supply and demand and plot for given runid.}
\usage{
hl_demandsupply(simp, runid = 0, dataname = "csv_cell_aggregated")
}
\arguments{
\item{simp}{}

\item{runid}{}

\item{dataname}{}
}
\value{
plot
}
\description{
Read supply and demand and plot for given runid.
}
\author{
Sascha Holzhauer
}

