% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_getAgentParamId}
\alias{hl_getAgentParamId}
\title{Fetch agent param ID from Runs.csv for paramId in given simp}
\usage{
hl_getAgentParamId(simp, agentParamColumn = "aftParamId",
  runIdColumn = "run")
}
\arguments{
\item{simp}{}

\item{agentParamColumn}{}

\item{runIdColumn}{}
}
\value{
agent Param ID
}
\description{
Fetch agent param ID from Runs.csv for paramId in given simp
}
\author{
Sascha Holzhauer
}

