% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_params.R
\name{hl_getAllocationRestrictionMatrix}
\alias{hl_getAllocationRestrictionMatrix}
\title{Returns matrix to define allocation restrictions.}
\usage{
hl_getAllocationRestrictionMatrix()
}
\value{
matrix with agent types as rows and columns
}
\description{
Fills the matrix with "0", meaning no restriction
}
\author{
Sascha Holzhauer
}

