% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_various.R
\name{hl_getBaseDirAdaptation}
\alias{hl_getBaseDirAdaptation}
\title{Extracts BaseDirAdaptation from Scenario.xml (empty string if not defined)}
\usage{
hl_getBaseDirAdaptation(simp)
}
\arguments{
\item{simp}{}
}
\value{
BaseDirAdaptation
}
\description{
Extracts BaseDirAdaptation from Scenario.xml (empty string if not defined)
}
\author{
Sascha Holzhauer
}
