% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_various.R
\name{hl_getCapitalDataFile}
\alias{hl_getCapitalDataFile}
\title{List required capital files for the given configuration}
\usage{
hl_getCapitalDataFile(simp, ID = NULL,
  cellInitialiserColName = "CellInitialisers", relativeToDataDir = TRUE)
}
\arguments{
\item{simp}{}

\item{ID}{}

\item{cellInitialiserColName}{}

\item{relativeToDataDir}{}
}
\value{
vector of filenames
}
\description{
List required capital files for the given configuration
}
\author{
Sascha Holzhauer
}
