% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_various.R
\name{hl_getCellNumbers}
\alias{hl_getCellNumbers}
\title{Retrieve information about the number of cells.}
\usage{
hl_getCellNumbers(simp, dataname = "csv_LandUseIndex_rbinded")
}
\arguments{
\item{simp}{}

\item{dataname}{}
}
\value{
list of data
}
\description{
Retrieve information about the number of cells.
}
\author{
Sascha Holzhauer
}

