% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_params.R
\name{hl_getCompetitionFunctions}
\alias{hl_getCompetitionFunctions}
\title{Read competition functions}
\usage{
hl_getCompetitionFunctions(simp, srcfilename = NULL,
  srcfilepath = dirname(paste(simp$dirs$output$data, simp$sim$folder,
  hl_getBaseDirAdaptation(simp), hl_getRunParameter(simp, "Competition_xml"),
  sep = "/")), runidcolumnname = "run", returntext = FALSE)
}
\arguments{
\item{simp}{}

\item{srcfilename}{competition XML file. Will be generated from \code{simp} if missing.}

\item{srcfilepath}{obtained from Links.csv by default. Assign \code{NULL} to apply
\code{simp$dirs$param$getparamdir}.}

\item{runidcolumnname}{}
}
\value{
functions
}
\description{
Read competition functions
}
\author{
Sascha Holzhauer
}

