% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_various.R
\name{hl_getRunInfo}
\alias{hl_getRunInfo}
\title{Return data in RunInfo table for runid defined in simp.
Considers \code{}}
\usage{
hl_getRunInfo(simp, filename = simp$dirs$output$runinfo)
}
\arguments{
\item{simp}{}

\item{filename}{}
}
\value{
data.frame
}
\description{
Return data in RunInfo table for runid defined in simp.
Considers \code{}
}
\author{
Sascha Holzhauer
}

