% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_various.R
\name{hl_getRunParameter}
\alias{hl_getRunParameter}
\title{Fetch agent param ID from Runs.csv for paramId in given simp}
\usage{
hl_getRunParameter(simp, parameter, runIdColumn = "run")
}
\arguments{
\item{simp}{}

\item{parameter}{parameter (colname) to fetch}

\item{runIdColumn}{}
}
\value{
parameter value for param ID defined in given \code{simp$sim$runids}.
}
\description{
Fetch agent param ID from Runs.csv for paramId in given simp
}
\author{
Sascha Holzhauer
}

