% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_gistatistics_singleAFT}
\alias{hl_gistatistics_singleAFT}
\title{Bar plot of number of takeovers per number of trials. Various regions as facets.}
\usage{
hl_gistatistics_singleAFT(simp, dataname = "csv_aggregateGiStatistics",
  regions = simp$sim$regions, facet_ncol = 1)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{facet_ncol}{}

\item{region}{}
}
\value{
plot
}
\description{
Bar plot of number of takeovers per number of trials. Various regions as facets.
}
\author{
Sascha Holzhauer
}

