% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_gistatistics_singleRegion}
\alias{hl_gistatistics_singleRegion}
\title{Bar plot of number of takeovers per number of trials}
\usage{
hl_gistatistics_singleRegion(simp, dataname = "csv_aggregateGiStatistics",
  regions = simp$sim$regions, facet_ncol = 1, numboxes = NULL,
  numberrange = NULL, trialrange = NULL)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{facet_ncol}{}

\item{numboxes}{number of bins}

\item{numberrange}{vector of two: considered range of numbers}

\item{trialrange}{vector of two: considered range of trials}

\item{region}{}
}
\value{
plot
}
\description{
Various AFTs as facets.
}
\author{
Sascha Holzhauer
}

