% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_landindiceskey_csv}
\alias{hl_landindiceskey_csv}
\title{Generates AFT key as CSV with columns 'Index' and 'LandUse'}
\usage{
hl_landindiceskey_csv(simp)
}
\arguments{
\item{simp}{\begin{itemize}
    \item \code{simp$mdata$aftNames}
 \item \code{simp$dirs$output$data}
    \end{itemize}}
}
\value{
csv file
}
\description{
Generates AFT key as CSV with columns 'Index' and 'LandUse'
}
\author{
Sascha Holzhauer
}

