% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_maps_aft}
\alias{hl_maps_aft}
\title{Plot land use map series}
\usage{
hl_maps_aft(simp, dataname = "raster_aft", ...)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{...}{passed to visualise_raster_printPlots}
}
\value{
map plot
}
\description{
Plot land use map series
}
\author{
Sascha Holzhauer
}

