% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_various.R
\name{hl_marginalutilities}
\alias{hl_marginalutilities}
\title{Reads marginal utilties and prepares the data for plotting}
\usage{
hl_marginalutilities(simp, filename = paste(simp$dirs$output$rdata,
  simp$sim$id, "Marginal_Utilities.csv", sep = "/"), storerdata = TRUE,
  returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{filename}{the file that contain marginal utility data. Default is \code{MarginalUtilities.csv} in RData folder.}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot
}
\description{
Reads marginal utilties and prepares the data for plotting
}
\author{
Sascha Holzhauer
}
