% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_normalisedutilities}
\alias{hl_normalisedutilities}
\title{Read and plot normalised per-cell residuals}
\usage{
hl_normalisedutilities(simp,
  filenamemarginalutils = paste(simp$dirs$output$rdata,
  "MarginalUtilitiesPerCell.csv", sep = "/"),
  filenamePlotPercellDemand = NULL,
  filenameNormalisedResiduals = paste("NormalisedResiduals", simp$sim$runid,
  sep = "_"), capitalfilepartorder = c("regionalisation", "U", "regions", "U",
  "datatype"))
}
\arguments{
\item{simp}{}

\item{filenamemarginalutils}{}

\item{filenamePlotPercellDemand}{}

\item{filenameNormalisedResiduals}{}

\item{capitalfilepartorder}{}
}
\value{
plot(s)
}
\description{
Read and plot normalised per-cell residuals
}
\author{
Sascha HolzhauerVar1
}

