% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_param_capital_map}
\alias{hl_param_capital_map}
\title{Create maps of initial capital levels}
\usage{
hl_param_capital_map(simp, capitals = simp$mdata$capitals,
  filenameorder = c("regionalisation", "U", "regions", "U", "datatype"))
}
\arguments{
\item{simp}{}

\item{capitals}{}

\item{filenameorder}{}
}
\value{
map plot
}
\description{
Create maps of initial capital levels
}
\author{
Sascha Holzhauer
}

