% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_params.R
\name{hl_plotAgentProductionParameters}
\alias{hl_plotAgentProductionParameters}
\title{Plot AFT's production as bar plot}
\usage{
hl_plotAgentProductionParameters(simp, filenameprefix = "AftProduction_",
  filenamepostfix = "_multi_medium", ggplotaddons = NULL,
  returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{filenameprefix}{}

\item{filenamepostfix}{}

\item{returnplot}{if true the ggplot object is returned}

\item{heading}{}
}
\value{
bar plot
}
\description{
Plot AFT's production as bar plot
}
\author{
Sascha Holzhauer
}
