% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_competitiveness.R
\name{hl_plotCompetitiveness}
\alias{hl_plotCompetitiveness}
\title{Calculate and plot competitiveness for every AFT on every cell}
\usage{
hl_plotCompetitiveness(simp, dirtocapitals = paste(simp$dirs$alloc,
  "/worlds/", simp$sim$world, "/regionalisations/", simp$sim$regionalisation,
  "/", simp$sim$scenario, sep = ""),
  capitalfilename = paste(simp$sim$regionalisation, "_", simp$sim$regions,
  "_Capitals.csv", sep = ""), dirtodemand = paste(simp$dirs$data, "/worlds/",
  simp$sim$world, "/regionalisations/", simp$sim$regionalisation, "/",
  simp$sim$scenario, sep = ""),
  demandfilename = paste(simp$sim$regionalisation, "_", simp$sim$scenario,
  "_", simp$sim$regions, "_demand.csv", sep = ""),
  dirtoproduction = paste(simp$dirs$data, "/production/", sep = ""),
  productionfilenamepattern = "<AFT>/AftProduction_<AFT>.csv",
  returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{dirtocapitals}{}

\item{capitalfilename}{}

\item{dirtodemand}{}

\item{demandfilename}{}

\item{dirtoproduction}{}

\item{productionfilenamepattern}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot
}
\description{
Calculate and plot competitiveness for every AFT on every cell
}
\author{
Sascha Holzhauer
}
