% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_printAgentProductionParameters}
\alias{hl_printAgentProductionParameters}
\title{Print AFT productivities}
\usage{
hl_printAgentProductionParameters(simp, filenameprefix = "AftProduction_",
  filenamepostfix = "_multi_medium", heading = "subsubsection")
}
\arguments{
\item{simp}{}

\item{filenameprefix}{}

\item{filenamepostfix}{}
}
\value{
LaTeX subsections and tables
}
\description{
Print AFT productivities
}
\author{
Sascha Holzhauer
}

