% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_printCompetitionFunctions}
\alias{hl_printCompetitionFunctions}
\title{Read and plot compotition functions}
\usage{
hl_printCompetitionFunctions(simp, srcfilename = NULL, xrange = c(-3, 3),
  yrange = c(-1, 1), filename = "competitionFunctions",
  runidcolumnname = "run")
}
\arguments{
\item{simp}{}

\item{srcfilename}{competition XML file}

\item{xrange}{Vector of two. The x range to plot}

\item{yrange}{Vector of two. The y range to plot}

\item{filename}{filename for figure}
}
\value{
plot
}
\description{
Read and plot compotition functions
}
\author{
Sascha Holzhauer
}

