% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_params.R
\name{hl_printRunParameters}
\alias{hl_printRunParameters}
\title{Print run parameters}
\usage{
hl_printRunParameters(simp)
}
\arguments{
\item{simp}{}

\item{filenameprefix}{}

\item{filenamepostfix}{}
}
\value{
print xtable
}
\description{
Print run parameters
}
\author{
Sascha Holzhauer
}

