% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_regions.R
\name{hl_regions_aftcomposition}
\alias{hl_regions_aftcomposition}
\title{Regional AFT composition}
\usage{
hl_regions_aftcomposition(simp, dataname = "csv_cell_aggregated",
  facet_ncol = 5)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{facet_ncol}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot
}
\description{
Regional AFT composition
}
\author{
Sascha Holzhauer
}

