% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_regions_demandandsupply}
\alias{hl_regions_demandandsupply}
\title{Regional demand and supply figure}
\usage{
hl_regions_demandandsupply(simp, runid = simp$sim$runids[1],
  dataname = "csv_cell_aggregated", facet_ncol = 5)
}
\arguments{
\item{simp}{}

\item{dataname}{}
}
\value{
plot
}
\description{
Regional demand and supply figure
}
\author{
Sascha Holzhauer
}

