% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_serviceproduction}
\alias{hl_serviceproduction}
\title{Load, aggregate and visualise production per service and AFT}
\usage{
hl_serviceproduction(simp, dataname = "csv_cell_aggregated", facet_ncol = 2,
  normaliseByAftNumber = TRUE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{facet_ncol}{}

\item{normaliseByAftNumber}{}
}
\value{
timeline plot
}
\description{
Load, aggregate and visualise production per service and AFT
}
\author{
Sascha Holzhauer
}

