% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_store_csvcelldata}
\alias{hl_store_csvcelldata}
\title{Store CSV cell data (also in splitted form)}
\usage{
hl_store_csvcelldata(simp, dataname = "csv_LandUseIndex", tickinterval = 10)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{tickinterval}{}
}
\description{
Store CSV cell data (also in splitted form)
}
\author{
Sascha Holzhauer
}

