% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_data_tables.R
\name{hl_table_afts}
\alias{hl_table_afts}
\title{Output aggregated number of AFTs as table with AFTs as columns}
\usage{
hl_table_afts(simp, dataname = "csv_aggregateAFTComposition",
  includeunmanaged = FALSE, aggcelldataname = "dataAgg", afts = NULL)
}
\arguments{
\item{simp}{}

\item{dataname}{name for retrieving aggregated AFT data (RData)}

\item{includeunmanaged}{included numbers for unmanaged cells if TRUE}

\item{aggcelldataname}{name for retrieving aggregated cell data (celldata) to calculate unmanagend}

\item{afts}{list of AFTs to limit set of shown AFTs}
}
\value{
data.frame or printed LaTeX table
}
\description{
Output aggregated number of AFTs as table with AFTs as columns
}
\author{
Sascha Holzhauer
}
