% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_takeovers_all}
\alias{hl_takeovers_all}
\title{Transition plot of AFT take overs (both due to giving in and giving up)}
\usage{
hl_takeovers_all(simp, runid = simp$sim$runids[1],
  landusedataname = "csv_LandUseIndex_rbinded",
  starttick = simp$sim$starttick, tickinterval = 10,
  endtick = simp$sim$endtick, datanametakeovers = "csv_aggregateTakeOver",
  dataname = "csv_cell_aggregated", grouping = c("Scenario", "Runid",
  "Region"))
}
\arguments{
\item{simp}{}

\item{runid}{}

\item{starttick}{}

\item{tickinterval}{}

\item{endtick}{}

\item{datanametakeovers}{}

\item{dataname}{}
}
\value{
transition plot
}
\description{
Transition plot of AFT take overs (both due to giving in and giving up)
}
\author{
Sascha Holzhauer
}

