% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_volatility}
\alias{hl_volatility}
\title{Read cell volatility data and visualise CellVolatility and NumVolatileCells as lines}
\usage{
hl_volatility(simp, dataname = "csv_aggregated_cellvolatility",
  datatype = "AggregateCellVolatility", datacolumns = c("CellVolatility",
  "NumVolatileCells"), percent = NULL, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{datatype}{}

\item{datacolumns}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot (and rData)
}
\description{
Read cell volatility data and visualise CellVolatility and NumVolatileCells as lines
}
\author{
Sascha Holzhauer
}

