% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_store.R
\name{hl_write_csv_landuseindex}
\alias{hl_write_csv_landuseindex}
\title{Outputs LandUseIndex data stored as RData object to CSV file(s).}
\usage{
hl_write_csv_landuseindex(simp, dataname = "csv_LandUseIndex_rbinded",
  filename = NULL, columns = c(simp$csv$cname_x, simp$csv$cname_y,
  "LandUseIndex"), pertick = TRUE, identifier = paste(simp$sim$version,
  simp$sim$scenario, sep = "_"), useIndices = TRUE, writeLegend = TRUE,
  zip = TRUE, removeCSV = zip, outdir = simp$dirs$output$reports,
  tempdir = NULL, ...)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{filename}{}

\item{columns}{}

\item{pertick}{}

\item{identifier}{part of filename}

\item{useIndices}{if FALSE, full AFT names will be written}

\item{writeLegend}{writes a separate CSV file containing the land use index legend}

\item{zip}{zip CSV file (apart from legend file)}

\item{removeCSV}{removes CSV file}

\item{tempdir}{if given, the CSV file will be stored here before zipped to outdir}
}
\value{
CSV file(s)
}
\description{
Considers \code{simp$sim$regions}.
}
\author{
Sascha Holzhauer
}

