% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_csv_data}
\alias{input_csv_data}
\title{Reads output data for the specified datatype amd dataname from CSV data for potentially multiple runs,
regions, and multiple ticks.}
\usage{
input_csv_data(simp, datatype = NULL, dataname = "Cell", colums = NULL,
  pertick = FALSE, attachfileinfo = TRUE)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{datatype}{(e.g. "Capital")}

\item{dataname}{(e.g. "Cap1")}

\item{colums}{vector of colum names. If given restricted returned colums to the given headers (plus X and Y coordinates)}

\item{pertick}{if TRUE the filename will be complemented by all available ticks}

\item{attachfileinfo}{}
}
\value{
list of data.frames containing requested data
}
\description{
Reads output data for the specified datatype amd dataname from CSV data for potentially multiple runs,
regions, and multiple ticks.
}
\author{
Sascha Holzhauer
}

