% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_csv_param_capitals}
\alias{input_csv_param_capitals}
\title{Reads capital levels for the specified capitals from CSV data for potentially multiple regions}
\usage{
input_csv_param_capitals(simp, capitals = simp$mdata$capitals)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{capitals}{vector of strings of requested capital levels}
}
\value{
list of data.frames containing capital levels
}
\description{
Reads capital levels for the specified capitals from CSV data for potentially multiple regions
}
\author{
Sascha Holzhauer
}

