% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_csv_param.R
\name{input_csv_param_capitals_cellnumbers}
\alias{input_csv_param_capitals_cellnumbers}
\title{Determines number of cells from CSV data for potentially multiple regions}
\usage{
input_csv_param_capitals_cellnumbers(simp, regionpartfromend = 2,
  regionpartdevider = "_")
}
\arguments{
\item{simp}{SIMulation Properties}

\item{regionpartfromend}{index of region part counted from end of filename (used to extract region from filename)}

\item{regionpartdevider}{character that separated the region filename parts (used to extract region from filename)}
}
\value{
data.frame with columns Region and Cells
}
\description{
TODO apply simp$sim$filepartordercapitals to extract region instead of regionpartfromend and regionpartdevider
}
\author{
Sascha Holzhauer
}

