% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_csv_param_demand}
\alias{input_csv_param_demand}
\title{Reads demand values from CSV data for potentially multiple regions}
\usage{
input_csv_param_demand(simp)
}
\arguments{
\item{simp}{SIMulation Properties}
}
\value{
list of data.frames containing demand values, and the filename of their origin
}
\description{
Reads demand values from CSV data for potentially multiple regions
}
\author{
Sascha Holzhauer
}

