% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_csv_param_productivities}
\alias{input_csv_param_productivities}
\title{Reads productivities for each capital and each service for the given aft}
\usage{
input_csv_param_productivities(simp, aft, filenameprefix = "AftProduction_",
  filenamepostfix = "_multi_medium")
}
\arguments{
\item{simp}{SIMulation Properties}

\item{aft}{}
}
\value{
data.frame containing productivities
}
\description{
Reads productivities for each capital and each service for the given aft
}
\author{
Sascha Holzhauer
}

