% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_csv_param_runs}
\alias{input_csv_param_runs}
\title{Read Runs data (Runs.csv)}
\usage{
input_csv_param_runs(simp)
}
\arguments{
\item{simp}{}
}
\value{
run data
}
\description{
Read Runs data (Runs.csv)
}
\author{
Sascha Holzhauer
}

