% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_csv_data.R
\name{input_csv_prealloccomp}
\alias{input_csv_prealloccomp}
\title{Read CSV data of pre-allocation competitiveness and transfer to data.frame of frequencies}
\usage{
input_csv_prealloccomp(simp, datatype = "PreAlloc", starttick = if
  (!is.null(simp$sim$starttick)) simp$sim$starttick else simp$tech$mintick,
  endtick = if (!is.null(simp$sim$endtick)) simp$sim$endtick else
  simp$tech$maxtick, tickinterval = simp$csv$tickinterval_agg)
}
\arguments{
\item{simp}{}

\item{datatype}{}

\item{starttick}{}

\item{endtick}{}

\item{tickinterval}{}
}
\value{
data.frame of frequencies per Tick, Comp and Above/Below column according 
frequencies
}
\description{
Read CSV data of pre-allocation competitiveness and transfer to data.frame of frequencies
}
\author{
Sascha Holzhauer
}
