% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_various.R
\name{input_processAftComposition}
\alias{input_processAftComposition}
\title{Proceses/Aggregates stored, aggregated AFT composistion data. Checks whether data contains absolute numbers or
proportions.}
\usage{
input_processAftComposition(simp, dataname = "csv_aggregateAFTComposition",
  aggregateRegions = TRUE)
}
\arguments{
\item{simp}{}

\item{dataname}{}
}
\value{
data.frame
}
\description{
Proceses/Aggregates stored, aggregated AFT composistion data. Checks whether data contains absolute numbers or
proportions.
}
\author{
Sascha Holzhauer
}
