% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_raster.R
\name{input_raster_get}
\alias{input_raster_get}
\title{Loads stored object for given id or read raster objects accordingly if not stored.}
\usage{
input_raster_get(simp, dataname = "raster_landUseIndex", storerdata = FALSE)
}
\arguments{
\item{simp}{}

\item{storerdata}{if \code{TRUE} If true, read raster data is stored in rData files. Note, this is not a good
idea when path changes, as the raster objects usually only store filenames instead of data values.}
}
\value{
list of raster objects
}
\description{
Loads stored object for given id or read raster objects accordingly if not stored.
}
\author{
Sascha Holzhauer
}
