% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_raster_output}
\alias{input_raster_output}
\title{Get list (regions) of lists (ticks) of raster data from files whose source is taken from simp.}
\usage{
input_raster_output(simp, datatype, dataname, starttick = 0,
  endtick = simp$tech$maxtick, tickinterval = 1, encapsefileinfo = TRUE)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{datatype}{category of raster data (e.g. "Capital")}

\item{dataname}{name of data of category (e.g. "Cap1")}

\item{starttick}{}

\item{endtick}{}

\item{tickinterval}{Note: tickinterval is based on starttick which defaults to 0!}
}
\value{
list of lists of raster data
}
\description{
Get list (regions) of lists (ticks) of raster data from files whose source is taken from simp.
}
\author{
Sascha Holzhauer
}

