% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_tools_buildsimplist}
\alias{input_tools_buildsimplist}
\title{Construct a list of SimP for given runs}
\usage{
input_tools_buildsimplist(runs, randomseed = 0)
}
\arguments{
\item{runs}{first part of runid}

\item{randomseed}{second part of runid}
}
\value{
list of SimPs
}
\description{
Construct a list of SimP for given runs
}
\author{
Sascha Holzhauer
}

