% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_checkexists}
\alias{input_tools_checkexists}
\title{Checks whether an object with the given object name has been stored under the given simp configuration}
\usage{
input_tools_checkexists(simp, objectName)
}
\arguments{
\item{simp}{}

\item{objectName}{}
}
\value{
TRUE if object already stored
}
\description{
Checks whether an object with the given object name has been stored under the given simp configuration
}
\author{
Sascha Holzhauer
}

