% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_constructFilenameList}
\alias{input_tools_constructFilenameList}
\title{Construct a list of filenames}
\usage{
input_tools_constructFilenameList(simp, datatype = NULL, dataname = NULL,
  order = simp$sim$filepartorder, considertick = TRUE)
}
\arguments{
\item{simp}{considered elements: \itemize{
\item{\code{simp$sim$filepartorder}}
\item{\code{simp$sim$scenario}}
\item{\code{simp$sim$regionalisation}}
\item{\code{simp$sim$runids}}
\item{\code{simp$sim$regions}}}}

\item{datatype}{}

\item{dataname}{}

\item{order}{}

\item{considertick}{}
}
\value{

}
\description{
Construct a list of filenames
}
\author{
Sascha Holzhauer
}

