% Generated by roxygen2 (4.0.1): do not edit by hand
\name{input_tools_getAvailableTicks}
\alias{input_tools_getAvailableTicks}
\title{Determine a vector of ticks for which the requested data (of type dataType and name dataName)
is available in dir. Assumes that all files in dir of the form *<datatype>-<dataname>-<tick>.<extension>
are requested. If extension is not given, any extension is considered. If datatype and dataname are
not given, the according parts of the filename is assumed not to exist (also the associated '-').}
\usage{
input_tools_getAvailableTicks(dir, datatype = NULL, dataname = NULL,
  extension = NULL, starttick = NULL, endtick = NULL,
  tickinterval = NULL)
}
\arguments{
\item{dir}{directory to search in}

\item{datatype}{(e.g. "Capital")}

\item{dataname}{(e.g. "Cap1")}

\item{extension}{used to restrict filename extension}

\item{starttick}{start tick of requested range of ticks}

\item{endtick}{end tick of requested range of ticks}

\item{tickinterval}{interval of requested range of ticks}
}
\value{
vector of ticks
}
\description{
Determine a vector of ticks for which the requested data (of type dataType and name dataName)
is available in dir. Assumes that all files in dir of the form *<datatype>-<dataname>-<tick>.<extension>
are requested. If extension is not given, any extension is considered. If datatype and dataname are
not given, the according parts of the filename is assumed not to exist (also the associated '-').
}
\author{
Sascha Holzhauer
}

