% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_getLinksValue}
\alias{input_tools_getLinksValue}
\title{Read a value from Links.cav according to simp and the given ID}
\usage{
input_tools_getLinksValue(simp, id = "CapitalFolder")
}
\arguments{
\item{simp}{}

\item{id}{}
}
\value{
value in Links.csv
}
\description{
Read a value from Links.cav according to simp and the given ID
}
\author{
Sascha Holzhauer
}

