% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_tools_getModelInputDir}
\alias{input_tools_getModelInputDir}
\title{Determines the model input folder for the given datadir}
\usage{
input_tools_getModelInputDir(simp, datatype)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{datatype}{one of c("capitals", "demand")}
}
\value{
model output directory as string
}
\description{
Determines the model input folder for the given datadir
}
\author{
Sascha Holzhauer
}

