% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_tools_getModelInputFilenames}
\alias{input_tools_getModelInputFilenames}
\title{Determines the model input filenames for the given simp and datatype, dataname, ticks settings.}
\usage{
input_tools_getModelInputFilenames(simp,
  folders = input_tools_getModelInputDir(simp, datatype), datatype = NULL,
  dataname = NULL, extension = NULL, pertick = FALSE, starttick = NULL,
  endtick = NULL, tickinterval = NULL)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{folders}{}

\item{datatype}{(e.g. "Capital")}

\item{dataname}{(e.g. "Cap1")}

\item{extension}{used to restrict filenames by extension (without "."!)}

\item{pertick}{if TRUE the filename will be complemented by all available ticks}

\item{starttick}{start tick of requested range of ticks}

\item{endtick}{end tick of requested range of ticks}

\item{tickinterval}{interval of requested range of ticks}
}
\value{
list of vector of filenames (list elements rerpesent files of one folder)
}
\description{
Determines the model input filenames for the given simp and datatype, dataname, ticks settings.
}
\author{
Sascha Holzhauer
}

