% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_getModelOutputFilenames}
\alias{input_tools_getModelOutputFilenames}
\title{Determines the model output filenames for the given simp and datatype, dataname, ticks settings}
\usage{
input_tools_getModelOutputFilenames(simp,
  folders = input_tools_getModelOutputDir(simp), datatype = NULL,
  dataname = NULL, extension = NULL, returnfileinfo = TRUE,
  pertick = FALSE, starttick = if (!is.null(simp$sim$starttick))
  simp$sim$starttick else simp$tech$mintick, endtick = if
  (!is.null(simp$sim$endtick)) simp$sim$endtick else simp$tech$maxtick,
  tickinterval = 1)
}
\arguments{
\item{simp}{}

\item{folders}{}

\item{datatype}{}

\item{dataname}{}

\item{extension}{}

\item{returnfileinfo}{}

\item{pertick}{if TRUE the filename will be complemented by all available ticks
(using \code{\link{input_tools_getAvailableTicks}})}

\item{starttick}{only required when \code{pertick == TRUE}}

\item{endtick}{only required when \code{pertick == TRUE}}

\item{tickinterval}{only required when \code{pertick == TRUE}}

\item{folders}{}
}
\value{
list of vector of filenames (list elements rerpesent files of one folder)
}
\description{
Able to read multiple scenarios (as in \code{simp$sim$scenario})/regions (as in \code{simp$sim$regions})/
runids (as in \code{simp$sim$runids}).
Attaches infos (region, runid, scenario).
}
\author{
Sascha Holzhauer
}
