% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_getParamValue}
\alias{input_tools_getParamValue}
\title{Rerieves param value for a param expression.
Consults Links.csv and Runs.csv as appropriate (recursively).}
\usage{
input_tools_getParamValue(simp, rawvalue)
}
\arguments{
\item{simp}{}

\item{rawvalue}{}
}
\value{
parameter value
}
\description{
Rerieves param value for a param expression.
Consults Links.csv and Runs.csv as appropriate (recursively).
}
\author{
Sascha Holzhauer
}

