% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_getrundata}
\alias{input_tools_getrundata}
\title{Get parameters in Runs.csv for the defined Run ID (in simp)}
\usage{
input_tools_getrundata(simp)
}
\arguments{
\item{simp}{}
}
\value{
data.frame of parameters in Runs.csv for the defined Run ID.
}
\description{
Get parameters in Runs.csv for the defined Run ID (in simp)
}
\author{
Sascha Holzhauer
}

