% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_tools_load}
\alias{input_tools_load}
\title{Wrapper for load}
\usage{
input_tools_load(simp, objectName, ...)
}
\arguments{
\item{simp}{\itemize{
    \item \code{simp$sim$id}
 \item \code{simp$dirs$output$data}
    \item \code{simp$sim$version} (\code{if is.null(simp$sim$id)}
}}

\item{objectName}{}
}
\description{
Wrapper for load
}
\author{
Sascha Holzhauer
}

