% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_loadstorecombined}
\alias{input_tools_loadstorecombined}
\title{Load, combine, and store data objects}
\usage{
input_tools_loadstorecombined(simp, simps, fromdataname,
  todataname = fromdataname)
}
\arguments{
\item{simp}{used for saving}

\item{simps}{list of simp}

\item{fromdataname}{character}

\item{todataname}{character}
}
\value{
stored RData object
}
\description{
Load, combine, and store data objects
}
\author{
Sascha Holzhauer
}

