% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_save}
\alias{input_tools_save}
\title{Wrapper for save}
\usage{
input_tools_save(simp, object)
}
\arguments{
\item{simp}{}

\item{object}{object name as character}
}
\description{
Wrapper for save
}
\author{
Sascha Holzhauer
}

