% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_xml_param_competition}
\alias{input_xml_param_competition}
\title{Reads competition functions for each service}
\usage{
input_xml_param_competition(simp, srcfilename = "Competition_linear")
}
\arguments{
\item{simp}{SIMulation Properties}

\item{aft}{}
}
\value{
data.frame containing productivities
}
\description{
Reads competition functions for each service
}
\author{
Sascha Holzhauer
}

