% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_agg.R
\name{metric_agg_actions_number}
\alias{metric_agg_actions_number}
\title{Get number of actions per institution(s)}
\usage{
metric_agg_actions_number(simp, institutions = NULL, pattern = NULL)
}
\arguments{
\item{simp}{}

\item{institutions}{}

\item{pattern}{if given, the action name must match the pattern (using grepl)}
}
\value{
number of actions
}
\description{
TODO test
}
\author{
Sascha Holzhauer
}

