% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_agg.R
\name{metric_agg_regionalsupply_efficiency}
\alias{metric_agg_regionalsupply_efficiency}
\title{Service delivery efficiency}
\usage{
metric_agg_regionalsupply_efficiency(simp, service = NULL,
  datanamesupply = "csv_aggregated_supply",
  datanameaft = "csv_aggregateAFTComposition",
  filenameprefix = "AftProduction_", filenamepostfix = "_multi_medium",
  filenameprefix_aftparams = "AftParams_", filenamepostfix_aftparams = "",
  aftwisefolder = TRUE, asvector = FALSE)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{service}{the service to calculate the metric for. If \code{NULL} the mean across services is returned.}

\item{datanamesupply}{}

\item{asvector}{if \code{TRUE} metric is returned as named vector}
}
\value{
data.frame with cols Metric, Ticks and Value or named vector (ticks) of metric when \code{asvector == TRUE}
}
\description{
Service delivery efficiency
}
\author{
Sascha Holzhauer
}

