% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_agg.R
\name{metric_agg_supplyaccrossreg_simpson}
\alias{metric_agg_supplyaccrossreg_simpson}
\title{TODO}
\usage{
metric_agg_supplyaccrossreg_simpson(simp, service = NULL,
  datanamesupply = "csv_aggregated_supply", asvector = FALSE)
}
\arguments{
\item{service}{the service to calculate the metric for. If \code{NULL} the mean across services is returned.}

\item{datanamesupply}{}

\item{asvector}{if \code{TRUE} metric is returned as named vector}
}
\value{
data.frame with cols Metric, Ticks and Value or named vector (ticks) of metric when \code{asvector == TRUE}
}
\description{
TODO
}
\author{
Sascha Holzhauer
}

