% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_agg.R
\name{metric_agg_supplydemand_percentage}
\alias{metric_agg_supplydemand_percentage}
\title{Calculates the percental supply relative to demand (supply = demand > 100) for all ticks}
\usage{
metric_agg_supplydemand_percentage(simp, service = "Total", regions = NULL,
  datanamedemand = "csv_aggregated_demand",
  datanamesupply = "csv_aggregated_supply", considerundersupply = TRUE,
  consideroversupply = FALSE, asvector = FALSE)
}
\arguments{
\item{simp}{}

\item{service}{character, the service the gap is returned for. "Total" for sum.}

\item{asvector}{if \code{TRUE} metric is returned as named vector}
}
\value{
data.frame with cols Metric, Ticks and Value or named vector (ticks) of metric (percental demand supply gap) when \code{asvector == TRUE}
}
\description{
Calculates the percental supply relative to demand (supply = demand > 100) for all ticks
}
\author{
Sascha Holzhauer
}
