% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_agg.R
\name{metric_aggaft_diversity_shannon}
\alias{metric_aggaft_diversity_shannon}
\title{Calculates the Shannon Diversity Index for AFT composition.}
\usage{
metric_aggaft_diversity_shannon(simp,
  dataname = "csv_aggregateAFTComposition", asvector = FALSE)
}
\arguments{
\item{simp}{Expects the defintion of one run}

\item{dataname}{}

\item{asvector}{if \code{TRUE} metric is returned as named vector}
}
\value{
data.frame with cols Metric, Ticks and Value or named vector (ticks) of metric (percental demand supply gap) when \code{asvector == TRUE}
}
\description{
TODO Consider \code{simp$sim$regions}
}
\author{
Sascha Holzhauer
}

