% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_agg.R
\name{metric_aggaft_diversity_simpson}
\alias{metric_aggaft_diversity_simpson}
\title{Regional land use diversity}
\usage{
metric_aggaft_diversity_simpson(simp, region = NULL,
  dataname = "csv_aggregateAFTComposition", asvector = FALSE)
}
\arguments{
\item{simp}{}

\item{region}{the region to calculate the metric for. If \code{NULL} the mean across regions is returned.}

\item{dataname}{}

\item{asvector}{if \code{TRUE} metric is returned as named vector}
}
\value{
data.frame with cols Metric, Ticks and Value or named vector (ticks) of metric when \code{asvector == TRUE}
}
\description{
Regional land use diversity
}
\author{
Sascha Holzhauer
}
