% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_agg.R
\name{metric_aggaft_proportions}
\alias{metric_aggaft_proportions}
\title{Return proportion of given set of AFTs of all agents (usually not unmanaged cells)}
\usage{
metric_aggaft_proportions(simp, afts, aftsname = NULL,
  dataname = "csv_aggregateAFTComposition", asvector = FALSE)
}
\arguments{
\item{simp}{}

\item{afts}{set (vector of chracters) of AFTs whose proportion is to be determined.}

\item{asvector}{if \code{TRUE} metric is returned as named vector}
}
\value{
data.frame with cols Metric, Ticks and Value or named vector (ticks) of metric when \code{asvector == TRUE}
}
\description{
Return proportion of given set of AFTs of all agents (usually not unmanaged cells)
}
\author{
Sascha Holzhauer
}

