% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_celldata.R
\name{metric_celldata_changes}
\alias{metric_celldata_changes}
\title{Determine number changes between ticks for a specific (combination of) land use(s)}
\usage{
metric_celldata_changes(simp, cdata, ncol = 1, title = "AFT-Changes",
  ggplotaddon = NULL, regions = simp$sim$regions,
  addcountryshapes = FALSE, plotunchanged = TRUE, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{cdata}{list of celldata (\code{data.frame)} to compare}

\item{ncol}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
vector of number of changes
}
\description{
Determine number changes between ticks for a specific (combination of) land use(s)
}
\author{
Sascha Holzhauer
}

