% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_raster.R
\name{metric_raster_global_patches}
\alias{metric_raster_global_patches}
\title{Determine the number of patches of a single raster}
\usage{
metric_raster_global_patches(simp, raster, directions = 8,
  relevantindices = NULL, relevantafts = NULL)
}
\arguments{
\item{simp}{}

\item{raster}{}

\item{directions}{}

\item{relevantindices}{raster values not included in \code{relevantindices} are considered 
to be background values and separate clumps.}

\item{relevantafts}{vector of AFT names. Alternative to \code{relevantindices} which has precedence.}
}
\value{
number of patches
}
\description{
Determine the number of patches of a single raster
}
\author{
Sascha Holzhauer
}

