% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_raster.R
\name{metric_rasters_changes}
\alias{metric_rasters_changes}
\title{Calculates the changes in land uses in a row of raster data.
TODO test
TODO limit to specific AFT(s)}
\usage{
metric_rasters_changes(simp, aft = NULL, dataname = "raster_landUseIndex",
  asvector = FALSE)
}
\arguments{
\item{aft}{if \code{NULL}, all AFTs are considered}

\item{asvector}{if \code{TRUE} metric is returned as named vector}
}
\value{
data.frame with cols Metric, Ticks and Value or named vector (ticks) of metric (changes (integer)) when \code{asvector == TRUE}
}
\description{
Calculates the changes in land uses in a row of raster data.
TODO test
TODO limit to specific AFT(s)
}
\author{
Sascha Holzhauer
}

