% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_regional_diversity}
\alias{metric_regional_diversity}
\title{TODO to complete
TODO think about un-regular regions (real-world) (use shapefiles?)}
\usage{
metric_regional_diversity(simp, data, regionwidth = 50,
  regionheight = regionwidth)
}
\arguments{
\item{simp}{}

\item{data}{}

\item{regionwidth}{}

\item{regionheight}{}
}
\value{
regional diversity metric
}
\description{
TODO to complete
TODO think about un-regular regions (real-world) (use shapefiles?)
}
\author{
Calum Brown

Sascha Holzhauer
}

