% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_raster.R
\name{metrics_rasters_connectivity}
\alias{metrics_rasters_connectivity}
\title{Calculates the connectivity (average of proportion of similar neighbours) in a series of raster data.
TODO test}
\usage{
metrics_rasters_connectivity(simp, afts = NULL,
  dataname = "raster_landUseIndex", asvector = FALSE)
}
\arguments{
\item{asvector}{if \code{TRUE} metric is returned as named vector}

\item{aft}{if \code{NULL}, all AFTs are considered}
}
\value{
data.frame with cols Metric, Ticks and Value or named vector (ticks) of metric (number of changed cells (integer)) when \code{asvector == TRUE}
}
\description{
Calculates the connectivity (average of proportion of similar neighbours) in a series of raster data.
TODO test
}
\author{
Sascha Holzhauer
}
