% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tools.R
\name{output_printList}
\alias{output_printList}
\title{Format vector of characters into latex list}
\usage{
output_printList(simp, textvector,
  out.format = knitr::opts_knit$get("out.format"), environment = "itemize",
  marker = NULL)
}
\arguments{
\item{simp}{}

\item{out.format}{}

\item{environment}{}

\item{marker}{}

\item{x}{}
}
\value{
latex formated list
}
\description{
Format vector of characters into latex list
}
\author{
Sascha Holzhauer
}
