% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tools.R
\name{output_tools_getDefaultFilename}
\alias{output_tools_getDefaultFilename}
\title{Create a default output filename of the pattern
<world>_<scenario>_<condensed runids>_<region>_<postifx>. Does not include extension!}
\usage{
output_tools_getDefaultFilename(simp, postfix = "")
}
\arguments{
\item{simp}{}

\item{postfix}{}
}
\value{
String of default output filename
}
\description{
Create a default output filename of the pattern
<world>_<scenario>_<condensed runids>_<region>_<postifx>. Does not include extension!
}
\author{
Sascha Holzhauer
}

