% Generated by roxygen2 (4.0.2): do not edit by hand
\name{output_visualise_aftFluctuations}
\alias{output_visualise_aftFluctuations}
\title{Visualise AFT fluctuations}
\usage{
output_visualise_aftFluctuations(simp, data, starttick = if
  (!is.null(simp$sim$starttick)) simp$sim$starttick else simp$tech$mintick,
  endtick = if (!is.null(simp$sim$endtick)) simp$sim$endtick else
  simp$tech$maxtick, tickinterval = 1, title = "AFT Fluctuations",
  filename = title)
}
\arguments{
\item{simp}{}

\item{data}{}

\item{starttick}{}

\item{endtick}{}

\item{tickinterval}{}

\item{title}{}

\item{filename}{}
}
\value{
plot
}
\description{
Visualise AFT fluctuations
}
\author{
Sascha Holzhauer
}

