% Generated by roxygen2 (4.0.1): do not edit by hand
\name{output_visualise_initFigure}
\alias{output_visualise_initFigure}
\title{Init figure output}
\usage{
output_visualise_initFigure(simp, outdir, filename, ensurePath = TRUE)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{outdir}{}

\item{filename}{}

\item{ensurePath}{switch}
}
\value{
plot
}
\description{
Init figure output
}
\author{
Sascha Holzhauer
}

