% Generated by roxygen2 (4.0.2): do not edit by hand
\name{output_visualise_takeovers}
\alias{output_visualise_takeovers}
\title{Visualise takeovers of land uses as transition plot (aka sankey diagram)}
\usage{
output_visualise_takeovers(simp, data, startpopulation = NULL,
  starttick = if (!is.null(simp$sim$starttick)) simp$sim$starttick else
  simp$tech$mintick, endtick = if (!is.null(simp$sim$endtick))
  simp$sim$endtick else simp$tech$maxtick, tickinterval = 1,
  type_of_arrow = "grid", transitionthreshold = 0,
  aftnames = simp$mdata$aftNames[-1])
}
\arguments{
\item{simp}{SIMulation Properties
\itemize{
    \item evtl. simp$sim$starttick or simp$tech$mintick
    \item evtl. simp$sim$endtick or simp$tech$maxtick
    \item simp$sim$runids
 \item simp$mdata$aftNames
    \item simp$dirs$output$figures
    \item simp$fig$init
    \item simp$sim$id
    \item simp$colours$AFT
    \item simp$fig$close
}}

\item{data}{takeover data with column "Tick" and one for each AFT (afts flow from row to column)}

\item{startpopulation}{needs to contain a column 'Number' which contains agent numbers of the start tick}

\item{starttick}{first tick to consider}

\item{endtick}{last tick to consider}

\item{tickinterval}{interval between transition columns. Note: tickinterval is based on starttick which defaults to 0!}

\item{type_of_arrow}{}

\item{transitionthreshold}{minimum number of transitions to show an arrow}
}
\value{
plots figure
}
\description{
In-between transitions are summed (in case tickinterval > 1), e.g. if single data points are given,
the transition Farmer (2010) > Forester (2011) > Farmer (2012) counts as two flows
Farmer (2010) > Forester (2011) and Forester (2011) > Farmer (2012).
}
\author{
Sascha Holzhauer
}
\seealso{
Other takeovers: \code{\link{convert_aggregate_takeovers}};
  \code{\link{hl_takeovers_all}};
  \code{\link{hl_takeovers}};
  \code{\link{output_visualise_aftFluctuations}}
}

