% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_getDefaultSimp}
\alias{param_getDefaultSimp}
\title{Get the default SIMP parameter list}
\usage{
param_getDefaultSimp()
}
\value{
list of default SIMP parameters
}
\description{
Get the default SIMP parameter list
}
\author{
Sascha Holzhauer
}

