% Generated by roxygen2 (4.0.2): do not edit by hand
\name{param_getExamplesSimp}
\alias{param_getExamplesSimp}
\title{Get the SIMP parameter list for working with example data}
\usage{
param_getExamplesSimp()
}
\value{
list of SIMP parameters to run examples
}
\description{
Get the SIMP parameter list for working with example data
}
\author{
Sascha Holzhauer
}

