% Generated by roxygen2 (4.0.2): do not edit by hand
\name{param_mergeDefaultSimp}
\alias{param_mergeDefaultSimp}
\title{Merge the given list into the default SIMP parameter list.
Applies \code{\link[utils]{mergeList}}.}
\usage{
param_mergeDefaultSimp(simp = list())
}
\arguments{
\item{simp}{}
}
\value{
given list modified by adding undefined parameter of SIMP default paramter list.
}
\description{
Merge the given list into the default SIMP parameter list.
Applies \code{\link[utils]{mergeList}}.
}
\author{
Sascha Holzhauer
}

