% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_raster.R
\name{read_raster_getMainFilename}
\alias{read_raster_getMainFilename}
\title{Concatenate the filename from basedir, scenario, world, run_id, year as well as datatype and dataname.
Region, datatype, and dataname are only considered if given as parameters.}
\usage{
read_raster_getMainFilename(basedir, scenario, world, run_id, region = NULL,
  dataType = NULL, dataName = NULL, year, useregiondir = FALSE)
}
\arguments{
\item{useregiondir}{(output data may or may not be stored in region specific folders)}

\item{datatype}{(e.g. "Capital")}

\item{dataname}{(e.g. "Cap1")}
}
\value{
String of filename
}
\description{
Concatenate the filename from basedir, scenario, world, run_id, year as well as datatype and dataname.
Region, datatype, and dataname are only considered if given as parameters.
}
\author{
Calum Brown

Sascha Holzhauer
}

