% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_raster_single}
\alias{read_raster_single}
\title{Read a single raster file for the given scenario/world/run_id/year combination in the given base dir.
Data type and data name a required to construct the filename properly.}
\usage{
read_raster_single(scenario = lix$scenario, world = pix$world,
  basedir = pix$path_names["dataSource"][[1]], run_id = lix$run_id,
  year = lix$year, region = NULL, dataType = NULL, dataName = NULL,
  useRegionDir = FALSE)
}
\arguments{
\item{scenario}{}

\item{world}{}

\item{basedir}{}

\item{run_id}{}

\item{year}{}

\item{dataType}{(e.g. "Capital")}

\item{dataName}{(e.g. "Cap1")}

\item{useRegionDir}{(output data may or may not be stored in region specific folders)}
}
\value{
rasterData
}
\description{
DEPRECATED!
}
\author{
Sascha Holzhauer
}

