% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_raster.R
\name{readingRasterData_collateRuns}
\alias{readingRasterData_collateRuns}
\title{Read raster of several years of several runIDs.}
\usage{
readingRasterData_collateRuns(scenario = lix$scenario, world = pix$world,
  basedir = pix$path_names["dataSource"][[1]], region = NULL,
  num_runs = NULL, dataType, dataName, aggregate = FALSE,
  runIDrange = NULL, useRegionDir = FALSE, func = NULL,
  startYear = NULL, endYear = NULL, everyYear = NULL)
}
\arguments{
\item{num_runs}{the number of runs to consider}

\item{useRegionDir}{(output data may or may not be stored in region specific folders)}
}
\value{
data.frame
}
\description{
deprecated? looks similar to input_raster_output but does not support SIMP
}
\author{
Sascha Holzhauer
}
