% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readingRasterData_collateRunsAndMeasures}
\alias{readingRasterData_collateRunsAndMeasures}
\title{Do title}
\usage{
readingRasterData_collateRunsAndMeasures(scenario = lix$scenario,
  world = pix$world, basedir = pix$path_names["dataSource"][[1]],
  region = NULL, num_runs = NULL, dataTypes, dataNames, aggregate = TRUE,
  runIDrange = NULL, useRegionDir = FALSE, funcs = NULL,
  startYear = lix$startTick, endYear = lix$endTick,
  everyYear = lix$everyTick)
}
\arguments{
\item{scenario}{}

\item{world}{}

\item{basedir}{}

\item{region}{}

\item{num_runs}{}

\item{dataTypes}{vector of dataTypes}

\item{dataNames}{according vector of dataName with same length as dataType}

\item{aggregate}{}

\item{runIDrange}{}

\item{useRegionDir}{(output data may or may not be stored in region specific folders)}

\item{endYear}{}

\item{everyYear}{}

\item{func}{}

\item{startyear}{}
}
\value{
data.frame
}
\description{
Do title
}
\author{
Sascha Holzhauer
}

