% Generated by roxygen2 (4.0.1): do not edit by hand
\name{settings_colours_getAftColours}
\alias{settings_colours_getAftColours}
\title{Return AFT colours}
\usage{
settings_colours_getAftColours()
}
\value{
coolour pallete
}
\description{
Return AFT colours
}
\author{
Sascha Holzhauer
}

