% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_maps.R
\name{visualisation_raster_legendandtitle}
\alias{visualisation_raster_legendandtitle}
\title{Ggplot2 theme that plots legend and title only}
\usage{
visualisation_raster_legendandtitle(base_size = 11,
  base_family = "Helvetica")
}
\arguments{
\item{base_size}{}

\item{base_family}{}
}
\value{
ggplot theme
}
\description{
Ggplot2 theme that plots legend and title only
}
\author{
Sascha Holzhauer
}

