% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visualisation_raster_nobackground}
\alias{visualisation_raster_nobackground}
\title{white background}
\usage{
visualisation_raster_nobackground(base_size = 11, base_family = "Helvetica")
}
\arguments{
\item{base_size}{}

\item{base_family}{}
}
\value{
theme
}
\description{
white background
}
\author{
Sascha Holzhauer
}

