% Generated by roxygen2 (4.0.1): do not edit by hand
\name{visualisation_raster_printRawPlots_theme_nothing}
\alias{visualisation_raster_printRawPlots_theme_nothing}
\title{GGplot2 theme for visualisation_raster_printRawPlots}
\usage{
visualisation_raster_printRawPlots_theme_nothing(base_size = 12,
  base_family = "Helvetica")
}
\arguments{
\item{base_size}{}

\item{base_family}{}
}
\value{
ggplot2 theme
}
\description{
GGplot2 theme for visualisation_raster_printRawPlots
}
\author{
Sascha Holzhauer
}

