% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_actions.R
\name{visualise_actions}
\alias{visualise_actions}
\title{Visualise the performance and effect of Actions}
\usage{
visualise_actions(simp, actiondata = "dataActions", onlyselected = FALSE,
  monitordata = NULL, title = NULL, filename = NULL, x_column = "Tick",
  y_column_action = "Score", y_column_measure = "Value",
  colour_column = "Service", shape_column = "Agent",
  fill_column = "Action", alpha_column = "Selected", alpha = 0.5,
  size_column = NULL, lineseparator_column_actions = NULL,
  linetype_column_measures = NULL, linetype_column_actions = NULL,
  facet_column = "Region", facet_ncol = 1, facet_column_levels = NULL,
  actionfillcolours = NULL, monitorcolours = NULL,
  actionlinecolour = "darkgrey", actionshapenumbers = c(21:25),
  returnplot = FALSE, ggplotaddons = NULL)
}
\arguments{
\item{simp}{}

\item{actiondata}{data.frame with columns \code{Action}, \code{Score}, \code{Selected}, \code{facet_column} and \code{x_column}.
If a character string, the data will be loaded.}

\item{onlyselected}{If \code{TRUE} only selected actions will be shown.}

\item{monitordata}{(melted) data.frame with columns \code{facet_column} and \code{x_column}.
If a character string, the data will be loaded. NOTE: the value for \code{facet_column} should be distinct from
the one in \code{actiondata} to show monitored data in a separate facet.}

\item{title}{figure title}

\item{filename}{filename for figure file}

\item{x_column}{X variable}

\item{y_column_action}{Y variable for action facets}

\item{y_column_measure}{Y variable for measure facets}

\item{colour_column}{Colour variable}

\item{shape_column}{Shape variable}

\item{fill_column}{d Fill variable for action facets}

\item{alpha_column}{Alpha variable}

\item{alpha}{lower alpha value (upper is 1)}

\item{size_column}{Size variable}

\item{lineseparator_column_actions}{used to draw separate line besides 'Agent'. A character string of comma-separated variables
is possible (feeds into \code{interaction)}.}

\item{facet_column}{Facet Variable}

\item{facet_ncol}{number of facet columns}

\item{facet_column_levels}{enables the user to specify the order of facets - vector of character strings.}

\item{monitorcolours}{named vector with keys for actiondata values in column \code{colour_column}}

\item{actionlinecolour}{Colour for lines connecting action symbols (character)}

\item{actionshapenumbers}{numbers of shapes for agents (action symbols)}

\item{returnplot}{if true the ggplot object is returned}

\item{ggplotaddons}{pass additional ggplot elements, e.g. \code{ggplot2::facet_grid(as.formula("Region ~ Runid"), scales="free_y")}.
Can be a single element or a list of elements.}

\item{linetype_column_measure}{Variable for linetype of measure facets}

\item{agentfillcolour}{named vector with keys for actiondata values in column \code{fill_column}}
}
\value{
figure
}
\description{
Visualise the performance and effect of Actions
}
\author{
Sascha Holzhauer
}
