% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visualise_bars}
\alias{visualise_bars}
\title{Prints a bar of data as ggplot2 with potentially different colour, linetype,
and potentially as facet plot.}
\usage{
visualise_bars(simp, data, y_column, title = NULL, fill_column = NULL,
  fill_legendtitle = fill_column, fill_legenditemnames = NULL,
  facet_column = NULL, facet_ncol = 4, filename = paste(title,
  shbasic::shbasic_condenseRunids(data[, "Runid"]), sep = "_"), alpha = 1,
  ggplotaddons = NULL, x_column = "ID", position = "dodge")
}
\arguments{
\item{simp}{SIMulation Properties}

\item{data}{data.frame or list that is rbinded to a data.frame}

\item{y_column}{}

\item{title}{figure title}

\item{fill_column}{column used to define colours}

\item{fill_legendtitle}{title for colour legend}

\item{fill_legenditemnames}{vector of names for colour legend items}

\item{facet_column}{column used to define facets}

\item{facet_ncol}{number of columns of facet wrap}

\item{filename}{without extension}

\item{alpha}{}

\item{ggplotaddons}{vector of ggplot objects to add}

\item{x_column}{}

\item{position}{passed to geom_bar}
}
\value{
ggplot2 line visualisation
}
\description{
Prints a bar of data as ggplot2 with potentially different colour, linetype,
and potentially as facet plot.
}
\examples{
#######################################################################
# crafty demo for csv data visualisation of AFT allocation
# Last update: 	27.10.2014
# Author: 		Sascha Holzhauer
# TODO adapt from rawPlot
#######################################################################
source(simp$simpDefinition)   # read simp
data <- input_csv_data(dataname = "AggregateAFTComposition")
visualisation_lines(data, legendTitle = "Capitals Example", factorial= FALSE, omitAxisTicks = FALSE, ncol = 1, id="None")
}
\author{
Sascha Holzhauer
}

