% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_cells.R
\name{visualise_cells_printRawPlots}
\alias{visualise_cells_printRawPlots}
\title{Prints a list of data.frames as raw ggplot2 facet plot.}
\usage{
visualise_cells_printRawPlots(simp, celldata, idcolumn = "Tick",
  valuecolumn = "LandUseIndex", title = "", filenamepostfix = title,
  factorial = FALSE, ncol = if (!is.data.frame(celldata)) length(celldata)
  else 1, coloursetname = simp$colours$defaultset, ggplotaddon = NULL,
  returnplot = FALSE)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{celldata}{(list of) data.frames contain info and X and X coordinates. If a list of data.frames,
elements must be named differently}

\item{idcolumn}{column used to separate and name rasters, refering to column names (set with colnames()) of the data.frame(s).}

\item{title}{name of plot}

\item{filenamepostfix}{appended to the default output filename @seealso output_tools_getDefaultFilename}

\item{factorial}{true if raster values are factorial (affects colour palette)}

\item{ncol}{number of columns of facet wrap. Defauls to the number of rasters in the first dataframe}

\item{coloursetname}{id for colour set (if factorial) to pass to simp$colours$GenericFun (e.g. "AFT", "Capital", "Service")}

\item{returnplot}{if true the ggplot object is returned}
}
\description{
There does not seem to be a straigh-forward way to convert a gTree object back to a ggplot2 object...
(http://stackoverflow.com/questions/29583849/r-saving-a-plot-in-an-object)
}
\author{
Sascha Holzhauer
}

