% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visualise_competition_prealloc}
\alias{visualise_competition_prealloc}
\title{Histogram of competitiveness per AFT: stacked: below/above GU threshold}
\usage{
visualise_competition_prealloc(simp, data,
  facet_ncol = length(simp$mdata$aftNames) - 1,
  filename = paste("PreAllocationCompetition", simp$sim$id, sep = "_"),
  numbins = 20,
  binwidth = diff(range(data$PreAllocCompetitiveness))/numbins,
  title = NULL, ggplotaddons = NULL, setfigdims = TRUE)
}
\arguments{
\item{simp}{}

\item{data}{}

\item{facet_ncol}{}

\item{filename}{}

\item{numbins}{number of bins to bin data}

\item{title}{plot title}

\item{ggplotaddons}{list of ggplot2 elements}

\item{setfigdims}{if \code{TRUE} \code{simp$fig$height} and \code{simp$fig$width} are set appropriately}
}
\value{
facet histogram plot
}
\description{
Histogram of competitiveness per AFT: stacked: below/above GU threshold
}
\author{
Sascha Holzhauer
}

