% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_competition.R
\name{visualise_competition_preallocTable}
\alias{visualise_competition_preallocTable}
\title{Histogram of competitiveness per AFT: stacked: below/above GU threshold}
\usage{
visualise_competition_preallocTable(simp, data,
  facet_ncol = length(simp$mdata$aftNames) - 1,
  filename = paste("PreAllocationCompetition", simp$sim$id, sep = "_"),
  numbins = 20, binwidth = NULL, title = NULL, ggplotaddons = NULL,
  setfigdims = TRUE, storename = NULL, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{data}{}

\item{facet_ncol}{}

\item{filename}{}

\item{numbins}{}

\item{binwidth}{}

\item{title}{}

\item{ggplotaddons}{}

\item{setfigdims}{}

\item{storename}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
bar plot
}
\description{
Histogram of competitiveness per AFT: stacked: below/above GU threshold
}
\author{
Sascha Holzhauer
}
