% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_lines.R
\name{visualise_lines}
\alias{visualise_lines}
\title{Prints a lines of data as ggplot2 with potentially different colour, linetype,
and potentially as facet plot.
Uses ticks as X variable if present, runid otherwise.}
\usage{
visualise_lines(simp, data, y_column, title = NULL, colour_column = NULL,
  colour_legendtitle = colour_column, colour_legenditemnames = NULL,
  linetype_column = NULL, linetype_legendtitle = linetype_column,
  linetype_legenditemnames = NULL, facet_column = NULL, facet_ncol = 2,
  filename = paste(title, shbasic::shbasic_condenseRunids(data[, "Runid"]),
  sep = "_"), alpha = simp$fig$alpha, ggplotaddons = NULL,
  returnplot = FALSE)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{data}{data.frame or list that is rbinded to a data.frame}

\item{title}{figure title}

\item{colour_column}{column used to define colours}

\item{colour_legendtitle}{title for colour legend}

\item{colour_legenditemnames}{vector of names for colour legend items}

\item{linetype_column}{column used to define linetypes}

\item{linetype_legendtitle}{title for linetype legend}

\item{linetype_legenditemnames}{vector of names for linetype legend items}

\item{facet_column}{column used to define facets}

\item{facet_ncol}{number of columns of facet wrap}

\item{filename}{without extension}

\item{ggplotaddons}{vector of ggplot objects to add}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
ggplot2 line visualisation
}
\description{
Prints a lines of data as ggplot2 with potentially different colour, linetype,
and potentially as facet plot.
Uses ticks as X variable if present, runid otherwise.
}
\examples{
#######################################################################
# crafty demo for csv data visualisation of AFT allocation
# Last update: 	27.10.2014
# Author: 		Sascha Holzhauer
# TODO adapt from rawPlot
#######################################################################
source(simp$simpDefinition)   # read simp
data <- input_csv_data(dataname = "AggregateAFTComposition")
visualisation_lines(data, legendTitle = "Capitals Example", factorial= FALSE, omitAxisTicks = FALSE, ncol = 1, id="None")
}
\author{
Sascha Holzhauer
}

