% Generated by roxygen2 (4.0.1): do not edit by hand
\name{visualise_raster_printPlots}
\alias{visualise_raster_printPlots}
\title{Prints a list of raster data as ggplot2 facet plot}
\usage{
visualise_raster_printPlots(simp, rasterdata, dataname = NULL,
  legendtitle = "", factorial = FALSE, omitaxisticks = FALSE, ncol = 1,
  coloursetname = "None", legenditemnames = NULL)
}
\arguments{
\item{rasterdata}{if a list, elements must be named differently}

\item{dataname}{vector of names of the same length as rasterData}

\item{legendtitle}{title for legend of raster values}

\item{factorial}{true if raster values are factorial (affects colour palette)}

\item{omitaxisticks}{omit axis ticks if true}

\item{ncol}{number of columns of facet wrap}

\item{coloursetname}{id for colour set (if factorial)}
}
\value{
raster visualisation
}
\description{
Prints a list of raster data as ggplot2 facet plot
}
\examples{
#######################################################################
# crafty demo for raster visualisation
# Last update: 	23.06.2014
# Author: 		Sascha Holzhauer
# TODO adapt from rawPlot
#######################################################################
source("demo/example_config.R")   # read simp
data <- input_csv_param_capitals()
raster <- convert_2raster(simp, data)
visualisation_raster_printRawPlots(raster, legendTitle = "Capitals Example", 
		factorial= FALSE, omitAxisTicks = FALSE, ncol = 1, id="None")
}
\author{
Sascha Holzhauer
}

