% Generated by roxygen2 (4.0.1): do not edit by hand
\name{visualise_raster_printRawPlots}
\alias{visualise_raster_printRawPlots}
\title{Writes a list of raster data as single raw (only the raster data) ggplot2 plot files}
\usage{
visualise_raster_printRawPlots(simp, rasterdata, datanames = NULL,
  legendtitle = "", factorial = FALSE, omitaxisticks = FALSE, ncol = 1,
  coloursetname = "None", legenditemnames = NULL)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{rasterdata}{if a list, elements must be named differently}

\item{datanames}{vector of names of the same length as rasterData}

\item{legendtitle}{title for legend of raster values}

\item{factorial}{true if raster values are factorial (affects colour palette)}

\item{omitaxisticks}{omit axis ticks if true}

\item{ncol}{number of columns of facet wrap}

\item{coloursetname}{name of colour set from settings_colours_getColorSet (if factorial)}
}
\value{
raster visualisation files
}
\description{
Writes a list of raster data as single raw (only the raster data) ggplot2 plot files
}
\examples{
#######################################################################
# crafty demo for raster visualisation
# Last update: 	23.06.2014
# Author: 		Sascha Holzhauer
#######################################################################
source(simp$simpDefinition)  # read simp
data <- input_csv_param_capitals()
raster <- convert_2raster(simp, data)
visualisation_raster_printRawPlots(raster, legendTitle = "Capitals Example", 
		factorial= FALSE, omitAxisTicks = FALSE, ncol = 1, id="None")
}
\author{
Sascha Holzhauer
}

