% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_2vectorfile}
\alias{convert_2vectorfile}
\title{Reads in CSV files with coordinates and creates polygone shapefiels via raster}
\usage{
convert_2vectorfile(indir, outdir, filenamesub = NULL)
}
\arguments{
\item{indir}{}

\item{outdir}{}

\item{filenamesub}{table with columns "number" and "sub" where the number must be contained in the original filename}
}
\value{
polygone shapefiles
}
\description{
Considers only CSV fiels in the indir. They must have a "x" and "y" column, but not "z"!
}
\author{
Sascha Holzhauer
}

