% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_aggregate_demand}
\alias{convert_aggregate_demand}
\title{Aggregate and store demand}
\usage{
convert_aggregate_demand(simp, checkexists = TRUE,
  demanddataname = "csv_aggregated_demand",
  sourcedataname = "dataAggregateSupplyDemand")
}
\arguments{
\item{simp}{}

\item{checkexists}{}

\item{demanddataname}{}
}
\value{
-
}
\description{
Tries to load demand from stored rData (usually from separate supply and demand CSV outputs).
If rData not present at \code{sourcedataname} it uses param data.
}
\author{
Sascha Holzhauer
}

