% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_aggregate_demand}
\alias{convert_aggregate_demand}
\title{Aggregate and store demand}
\usage{
convert_aggregate_demand(simp, checkexists = TRUE,
  demanddataname = "csv_aggregated_demand")
}
\arguments{
\item{simp}{}

\item{checkexists}{}

\item{demanddataname}{}
}
\value{
-
}
\description{
Aggregate and store demand
}
\author{
Sascha Holzhauer
}

