% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_aggregate.R
\name{convert_aggregate_supply}
\alias{convert_aggregate_supply}
\title{Aggregate and store supply data}
\usage{
convert_aggregate_supply(simp, celldataname = "csv_cell_aggregated",
  checkexists = TRUE, supplydataname = "csv_aggregated_supply",
  sourcedataname = "dataAggregateSupplyDemand")
}
\arguments{
\item{simp}{considered elements: \itemize{
\item{\code{simp$mdata$conversion$services}}
\item{\code{\link{input_tools_load}}}}}

\item{celldataname}{aggregated cell CSV data}

\item{checkexists}{}

\item{supplydataname}{name for storing supply data}

\item{sourcedataname}{name for retrieving data (rData)}
}
\value{
-
}
\description{
Tries to load supply from stored rData (usually from separate supply and demand CSV outputs).
If rData not present at \code{sourcedataname} it uses cell data.
}
\author{
Sascha Holzhauer
}

