% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2shapefile.R
\name{convert_csv2shapefile}
\alias{convert_csv2shapefile}
\title{Convert X and Y coordinates from a CSV file into a shapefile}
\usage{
convert_csv2shapefile(sip, indir, outdir,
  countryCodesFile = system.file("extdata",
  "various/CountryCodeNumberMapping.csv", package = "crafty"))
}
\arguments{
\item{sip}{}

\item{indir}{}

\item{outdir}{}

\item{countryCodesFile}{}
}
\value{
shapefile
}
\description{
NOTE: not tested yet
}
\author{
Sascha Holzhauer
}

