% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_raster_flatlist}
\alias{convert_raster_flatlist}
\title{Aggregate raster data with the given aggregate function}
\usage{
convert_raster_flatlist(inforasterdata, aggregatefunction = "sum")
}
\arguments{
\item{inforasterdata}{}

\item{aggregatefunction}{}
}
\value{
data.frame
}
\description{
Aggregate raster data with the given aggregate function
}
\author{
Sascha Holzhauer
}

