% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_raster_aggregate.R
\name{convert_raster_getAftNumbers}
\alias{convert_raster_getAftNumbers}
\title{Compute AFt numbers}
\usage{
convert_raster_getAftNumbers(aft_raster)
}
\arguments{
\item{aft_raster}{}
}
\value{
data.farme with numbers and info
}
\description{
Compute AFt numbers
}
\author{
Sascha Holzhauer
}

