% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{crafty}
\alias{crafty}
\alias{crafty-package}
\title{Processing CRAFTY model output}
\description{
The package offers methods to aggregate, process, and output data from CRAFTY simulations.
}
\details{
Postprocessing for CRAFTY simulations

\tabular{ll}{
Package \tab		crafty \cr
Model:  \tab		<Basic> \cr
Licence:\tab		GPL v3 \cr
Be aware of:\tab	requires libraries ggplot2, RColorBrewer, R.oo \cr
Version: \tab	0.1.0 \cr
Date	   	\tab 	2014-06-27 \cr
Changes	\tab	refactoring, comment \cr
}
}
\examples{
source(simp$simpDefinition)
input_csv_param_capitals()
}
\author{
Sascha Holzhauer, CESR/UoE \email{Sascha.Holzhauer@ed.ac.uk},
			Calum Brown, UoE, \email{Calum.Brown@ed.ac.uk}
}
\references{
TODO
}
\keyword{ABM}
\keyword{LUCC}
\keyword{change}
\keyword{crafty}
\keyword{land}
\keyword{post-processing}
\keyword{use}

