% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_xml_param.R
\name{get_xmlfunction_parameter}
\alias{get_xmlfunction_parameter}
\title{Extract parameter from source XML file}
\usage{
get_xmlfunction_parameter(data, parameter, default)
}
\arguments{
\item{data}{data.frame}

\item{parameter}{}

\item{default}{}

\item{rundata}{}
}
\value{
parameter value
}
\description{
Considers given default value in case the parameter is not defined in the XML file.
}
\author{
Sascha Holzhauer Links.csv
}

