% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_aftcomposisition_file}
\alias{hl_aftcomposisition_file}
\title{Reads single CSV file with AFT composition and produces timeline figure}
\usage{
hl_aftcomposisition_file(simp, csvfilename, title = "AftComposition",
  figurefilename = "AftComposition", returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{csvfilename}{}

\item{title}{}

\item{figurefilename}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
figure
}
\description{
Reads single CSV file with AFT composition and produces timeline figure
}
\author{
Sascha Holzhauer
}
