% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_aftcomposition}
\alias{hl_aftcomposition}
\title{Load, aggregate and visualise AFT composition data}
\usage{
hl_aftcomposition(simp, dataname = "csv_cell_aggregated")
}
\arguments{
\item{simp}{}

\item{dataname}{}
}
\value{
timeline plot
}
\description{
Load, aggregate and visualise AFT composition data
}
\author{
Sascha Holzhauer
}

