% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_aftcomposition}
\alias{hl_aftcomposition}
\title{Load, aggregate and visualise AFT composition data}
\usage{
hl_aftcomposition(simp, dataname = "csv_cell_aggregated",
  returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{returnplot}{if true, return ggplot object}
}
\value{
timeline plot
}
\description{
Usually, does not include the number of unmanaged cells.
}
\author{
Sascha Holzhauer
}
