% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_aftmap}
\alias{hl_aftmap}
\title{AFT map for 2010 and 2040 with differences from stored CSV data}
\usage{
hl_aftmap(simp, dataname = "csv_LandUseIndex_rbinded", firsttick = 2010,
  secondtick = 2040, ncol = 1, ggplotaddon = NULL)
}
\arguments{
\item{simp}{}

\item{dataname}{name of stored CSV LandUseIndex data}
}
\value{
map plot
}
\description{
AFT map for 2010 and 2040 with differences from stored CSV data
}
\author{
Sascha Holzhauer
}

