% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_maps.R
\name{hl_aftmap_changes}
\alias{hl_aftmap_changes}
\title{Plot changes between ticks for a specific (combination of) land use(s)}
\usage{
hl_aftmap_changes(simp, cdata, ncol = 1, title = "AFT-Changes",
  ggplotaddon = NULL, regions = simp$sim$regions,
  addcountryshapes = FALSE, plotunchanged = TRUE, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{cdata}{list of celldata (\code{data.frame)} to compare}

\item{ncol}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
figure file / plot
}
\description{
Plot changes between ticks for a specific (combination of) land use(s)
}
\author{
Sascha Holzhauer
}
