% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_aftmap_changes_temporal}
\alias{hl_aftmap_changes_temporal}
\title{Plot changes between ticks for a specific (combination of) land use(s)}
\usage{
hl_aftmap_changes_temporal(simp, dataname = "csv_LandUseIndex_rbinded",
  selectedAFT = 1, regions = simp$sim$regions, starttick = 2010,
  endtick = 2040, ncol = 1, title = "AFT-Changes", ggplotaddon = NULL,
  addcountryshapes = FALSE, plotunchanged = TRUE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{selectedAFT}{}

\item{starttick}{}

\item{endtick}{}

\item{ncol}{}

\item{ggplotaddon}{}
}
\value{
figure file / plot
}
\description{
Plot changes between ticks for a specific (combination of) land use(s)
}
\author{
Sascha Holzhauer
}

