% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_afttakeoverfluctuations}
\alias{hl_afttakeoverfluctuations}
\title{AFT take over fluctations as timeline}
\usage{
hl_afttakeoverfluctuations(simp, dataname = "csv_cell_aggregated",
  starttick = 2010, tickinterval = 5, endtick = 2040,
  datanametakeovers = "csv_aggregateTakeOver")
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{starttick}{}

\item{tickinterval}{}

\item{endtick}{}

\item{datanametakeovers}{}

\item{runid}{}
}
\value{
timeline plot
}
\description{
AFT take over fluctations as timeline
}
\author{
Sascha Holzhauer
}

