% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_aggregate_aftcompositions}
\alias{hl_aggregate_aftcompositions}
\title{Yearly aggregated AFT composition}
\usage{
hl_aggregate_aftcompositions(simp, dataname = "csv_aggregateAFTComposition",
  includeunmanaged = FALSE, aggcelldataname = "dataAgg",
  returnplot = FALSE, afts = NULL)
}
\arguments{
\item{simp}{}

\item{includeunmanaged}{if \code{TRUE} the number of unmanaged cells is plotted, too. It is derived from the
total number of cells obtained}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
timelien plot
}
\description{
Yearly aggregated AFT composition
}
\seealso{
hl_comp_cell_aftcomposition
}
\author{
Sascha Holzhauer
}
