% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_aggregate_demandsupply}
\alias{hl_aggregate_demandsupply}
\title{Yearly aggregated demand and supply}
\usage{
hl_aggregate_demandsupply(simp, dataname = "csv_aggregateServiceDemand",
  title = "Aggregated Service Supply & Demand", returnplot = FALSE,
  services = NULL)
}
\arguments{
\item{simp}{}

\item{returnplot}{if true the ggplot object is returned}

\item{services}{vector of services labels to visualise. If NULL, all services are displayed}
}
\value{
timeline plot
}
\description{
Yearly aggregated demand and supply
}
\author{
Sascha Holzhauer
}

