% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_csv.R
\name{hl_aggregate_demandsupply_csv}
\alias{hl_aggregate_demandsupply_csv}
\title{Aggregate demand and supply for all regions per tick and store as CSV file}
\usage{
hl_aggregate_demandsupply_csv(simp, dataname = "dataAggregateSupplyDemand")
}
\arguments{
\item{simp}{}
}
\value{
csv file
}
\description{
Aggregate demand and supply for all regions per tick and store as CSV file
}
\author{
Sascha Holzhauer
}
