% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_comp_aggregate_aftcompositions}
\alias{hl_comp_aggregate_aftcompositions}
\title{Compare Yearly aggregated AFT composition}
\usage{
hl_comp_aggregate_aftcompositions(simp, simps,
  dataname = "csv_aggregateAFTComposition",
  filename = paste("AftComposition_", if (!is.null(simp$sim$rundesc))
  paste(simp$sim$rundesc, collapse = "-") else simp$sim$id, sep = ""),
  title = "Aft Composition")
}
\arguments{
\item{simp}{}

\item{dataname}{}
}
\value{
timelien plot
}
\description{
Compare Yearly aggregated AFT composition
}
\author{
Sascha Holzhauer
}

