% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_comp_aggregate.R
\name{hl_comp_aggregate_aftcompositions}
\alias{hl_comp_aggregate_aftcompositions}
\title{Compare Yearly aggregated AFT composition}
\usage{
hl_comp_aggregate_aftcompositions(simp, simps,
  dataname = "csv_aggregateAFTComposition",
  filename = paste("AftComposition_", if (!is.null(simp$sim$rundesc))
  paste(simp$sim$rundesc, collapse = "-") else simp$sim$id, sep = ""),
  title = "Aft Composition", returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
timelien plot
}
\description{
Compare Yearly aggregated AFT composition
}
\author{
Sascha Holzhauer
}

