% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_comp_aggregate.R
\name{hl_comp_cell_aftcomposition}
\alias{hl_comp_cell_aftcomposition}
\title{Load, aggregate and visualise AFT composition data of several runs to compare}
\usage{
hl_comp_cell_aftcomposition(simp, simps, dataname = "csv_cell_aggregated",
  returnplot = FALSE)
}
\arguments{
\item{simps}{list of simp that shall be compared. See \code{\link{input_tools_load}} for relevant entries in
\code{simp}). \code{simp$sim$shortid} is combined with \code{Runid} in data to distinguish data!}

\item{dataname}{name of aggregated data from cell data}

\item{returnplot}{if true the ggplot object is returned}

\item{title}{title for plot}
}
\value{
timeline plot
}
\description{
Load, aggregate and visualise AFT composition data of several runs to compare
}
\author{
Sascha Holzhauer
}

