% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_comp_demandsupply}
\alias{hl_comp_demandsupply}
\title{Read supply and demand and plot for given runid.}
\usage{
hl_comp_demandsupply(simp, simps, dataname = "csv_cell_aggregated")
}
\arguments{
\item{simp}{}

\item{dataname}{aggregated demand and supply (from aggregated CSV)}

\item{runid}{}
}
\value{
plot
}
\description{
Read supply and demand and plot for given runid.
}
\author{
Sascha Holzhauer
}

