% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_comp_aggregate.R
\name{hl_comp_demandsupply}
\alias{hl_comp_demandsupply}
\title{Read supply and demand and plot for given runid}
\usage{
hl_comp_demandsupply(simp, simps, dataname = "csv_cell_aggregated",
  filename = paste("TotalDemandAndSupply_", if (!is.null(simp$sim$rundesc))
  paste(simp$sim$rundesc, collapse = "-") else simp$sim$id, sep = ""),
  title = paste("Demand & Supply (", paste(simp$sim$rundesc, collapse = "/"),
  ")", sep = ""), visualise = TRUE)
}
\arguments{
\item{simps}{list of simp that shall be compared. Relevant entries in
\code{simp}:\itemize{
    \item{\code{p$sim$shortid}}
    \item{\code{\link{convert_aggregate_demand}}}
    \item{\code{\link{convert_aggregate_supply}}}
    \item{\code{\link{input_tools_load}}}}
\code{simp$sim$shortid} is combined with \code{Runid} in data to distinguish data!}

\item{dataname}{aggregated demand and supply (from aggregated CSV)}

\item{title}{title for plot}

\item{visualise}{if \code{FALSE}, only combined and aggregated demand and supply data is returned}

\item{runid}{}
}
\value{
plot, combined supply and demand data
}
\description{
Change IDs to meaningful names by setting \code{simp$sim$shortid} for \code{simps}
}
\author{
Sascha Holzhauer
}

