% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_competitiveness}
\alias{hl_competitiveness}
\title{Load from csv data, aggregate, and plot AFT competitiveness over time (only for ticks cell data is available for)}
\usage{
hl_competitiveness(simp, dataname = "csv_cell_aggregated",
  returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot
}
\description{
Load from csv data, aggregate, and plot AFT competitiveness over time (only for ticks cell data is available for)
}
\author{
Sascha Holzhauer
}

