% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_compileruninfos}
\alias{hl_compileruninfos}
\title{Print LaTeX table including run information}
\usage{
hl_compileruninfos(simp, filename = simp$dirs$output$runinfo, rows = NULL)
}
\arguments{
\item{simp}{}

\item{filename}{}

\item{rows}{}
}
\value{
xtable plot
}
\description{
Considers the version in \code{simp$sim$version} and
the '1st Run ID' in \code{simp$sim$runids[1]}.
}
\author{
Sascha Holzhauer
}

