% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_compileruninfos}
\alias{hl_compileruninfos}
\title{Print LaTeX table including run information for the version in simp$sim$version.}
\usage{
hl_compileruninfos(simp, filename = simp$dirs$output$runinfo, rows = NULL)
}
\arguments{
\item{simp}{}

\item{filename}{}

\item{rows}{}
}
\value{
xtable plot
}
\description{
Print LaTeX table including run information for the version in simp$sim$version.
}
\author{
Sascha Holzhauer
}

