% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_params.R
\name{hl_compilerunparams}
\alias{hl_compilerunparams}
\title{Print LaTeX table including run parameters}
\usage{
hl_compilerunparams(simp, runidcolumnname = "run")
}
\arguments{
\item{simp}{}
}
\value{
xtable plot
}
\description{
Considers the \code{simp$sim$version} in column run.
}
\author{
Sascha Holzhauer
}
\seealso{
input_csv_param_runs
}

