% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_connectedness}
\alias{hl_connectedness}
\title{Visualise stored metric LandUseConnectivity as timelines. Stores RData from CSV if not done yet.}
\usage{
hl_connectedness(simp, dataname = "csv_aggregated_connectivity",
  datatype = "LandUseConnectivity", aftcolumns = simp$mdata$aftNames[-1],
  percent = NULL, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{datatype}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot (and rData)
}
\description{
Visualise stored metric LandUseConnectivity as timelines. Stores RData from CSV if not done yet.
}
\author{
Sascha Holzhauer
}
