% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_demandsupply}
\alias{hl_demandsupply}
\title{Read supply and demand and plot for given runid}
\usage{
hl_demandsupply(simp, runid = simp$sim$runids,
  dataname = "csv_cell_aggregated", returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{runid}{}

\item{dataname}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot
}
\description{
Read supply and demand and plot for given runid
}
\author{
Sascha Holzhauer
}
