% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_gistatistics_singleAFT}
\alias{hl_gistatistics_singleAFT}
\title{Bar plot of number of takeovers per number of trials}
\usage{
hl_gistatistics_singleAFT(simp, dataname = "csv_aggregateGiStatistics",
  regions = simp$sim$regions, facet_ncol = 1, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{returnplot}{if true the ggplot object is returned}

\item{region}{}
}
\value{
plot
}
\description{
Various regions as facets.
}
\author{
Sascha Holzhauer
}

