% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_gistatistics_singleRegion}
\alias{hl_gistatistics_singleRegion}
\title{Bar plot of number of takeovers per number of trials}
\usage{
hl_gistatistics_singleRegion(simp, dataname = "csv_aggregateGiStatistics",
  regions = simp$sim$regions, facet_ncol = 1, numboxes = NULL,
  numberrange = c(-Inf, Inf), trialrange = NULL, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{numboxes}{number of bins}

\item{numberrange}{vector of two: considered range of numbers}

\item{trialrange}{vector of two: considered range of trials}

\item{returnplot}{if true the ggplot object is returned}

\item{region}{}
}
\value{
plot
}
\description{
Various AFTs as facets.
}
\author{
Sascha Holzhauer
}

