% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_various.R
\name{hl_landindiceskey_csv}
\alias{hl_landindiceskey_csv}
\title{Generates AFT key as CSV with columns 'Index' and 'LandUse'}
\usage{
hl_landindiceskey_csv(simp)
}
\arguments{
\item{simp}{\itemize{
    \item \code{simp$mdata$aftNames}
 \item \code{simp$dirs$output$data}
}}
}
\value{
csv file
}
\description{
Generates AFT key as CSV with columns 'Index' and 'LandUse'
}
\author{
Sascha Holzhauer
}

