% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_lines_from_csv}
\alias{hl_lines_from_csv}
\title{Read data and visualise datacolumns as lines}
\usage{
hl_lines_from_csv(simp, dataname, datatype, datacolumns = NULL,
  linetypecol = "ID", colourcol = "Type", titleprefix = NULL,
  filenameprefix = NULL, percent = NULL)
}
\arguments{
\item{simp}{}

\item{dataname}{name of rData object}

\item{datatype}{datatype of CSV data (only relevant if rData not yet existing)}

\item{datacolumns}{columns to plot}

\item{linetypecol}{}

\item{colourcol}{}

\item{titleprefix}{}

\item{filenameprefix}{}

\item{percent}{data.frame with frist column containing the grouping (and named with the grouping variable)
the 2nd column values shall be considered as 100\% for}
}
\value{
plot (and rData)
}
\description{
Reads data from CSV file if not existing as rData object and visualises datacolums.
}
\author{
Sascha Holzhauer
}

