% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_rastermaps.R
\name{hl_maps_aft}
\alias{hl_maps_aft}
\title{Plot land use map series}
\usage{
hl_maps_aft(simp, dataname = "raster_aft", ncol = 5, returnplot = FALSE,
  ...)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{returnplot}{if true the ggplot object is returned}

\item{...}{passed to visualise_raster_printPlots}
}
\value{
map plot
}
\description{
Plot land use map series
}
\author{
Sascha Holzhauer
}
