% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_marginalutilities}
\alias{hl_marginalutilities}
\title{Reads marginal utilties and prepares the data for plotting}
\usage{
hl_marginalutilities(simp, filename = paste(simp$dirs$output$rdata,
  "MarginalUtilities.csv", sep = "/"), storerdata = TRUE)
}
\arguments{
\item{simp}{}

\item{filename}{}
}
\value{
data.frame
}
\description{
Stores data and returns.
}
\author{
Sascha Holzhauer
}

