% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_various.R
\name{hl_marginalutilities}
\alias{hl_marginalutilities}
\title{Reads marginal utilties and prepares the data for plotting}
\usage{
hl_marginalutilities(simp, filename = paste(simp$dirs$output$rdata,
  "MarginalUtilities.csv", sep = "/"), storerdata = TRUE,
  returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{filename}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
data.frame
}
\description{
Stores data and returns.
}
\author{
Sascha Holzhauer
}

