% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_normalisedutilities}
\alias{hl_normalisedutilities}
\title{Read and plot normalised per-cell residuals}
\usage{
hl_normalisedutilities(simp,
  filenamemarginalutils = paste(simp$dirs$output$rdata,
  "MarginalUtilitiesPerCell.csv", sep = "/"),
  filenamePlotPercellDemand = NULL,
  filenameNormalisedResiduals = paste("NormalisedResiduals", simp$sim$runid,
  sep = "_"), capitalfilepartorder = c("regionalisation", "U", "regions", "U",
  "datatype"), returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{filenamemarginalutils}{}

\item{filenamePlotPercellDemand}{}

\item{filenameNormalisedResiduals}{}

\item{capitalfilepartorder}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot(s)
}
\description{
Read and plot normalised per-cell residuals
}
\author{
Sascha HolzhauerVar1
}

