% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_param_maps.R
\name{hl_param_capital_map}
\alias{hl_param_capital_map}
\title{Create maps of initial capital levels}
\usage{
hl_param_capital_map(simp, capitals = simp$mdata$capitals,
  filenameorder = c("regionalisation", "U", "regions", "U", "datatype"),
  returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{capitals}{}

\item{filenameorder}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
map plot
}
\description{
Create maps of initial capital levels
}
\author{
Sascha Holzhauer
}

