% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_params.R
\name{hl_printAgentParameters}
\alias{hl_printAgentParameters}
\title{Print agent parameters}
\usage{
hl_printAgentParameters(simp, filenameprefix = "AftParams_",
  filenamepostfix = "", aftParamId = NULL, columnindices = c(2:7, 9),
  returnDataFrame = FALSE)
}
\arguments{
\item{simp}{}

\item{filenameprefix}{}

\item{filenamepostfix}{}

\item{if}{\code{TRUE} no LaTeX table is printed but the data.frame returned.}
}
\value{
print xtable
}
\description{
Print agent parameters
}
\author{
Sascha Holzhauer
}
