% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_printAgentParameters}
\alias{hl_printAgentParameters}
\title{Print agent parameters}
\usage{
hl_printAgentParameters(simp, filenameprefix = "AftParams_",
  filenamepostfix = "", aftParamId = 0)
}
\arguments{
\item{simp}{}

\item{filenameprefix}{}

\item{filenamepostfix}{}

\item{aftParamId}{}
}
\value{
print xtable
}
\description{
Print agent parameters
}
\author{
Sascha Holzhauer
}

