% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_params.R
\name{hl_printAgentProductionParameters}
\alias{hl_printAgentProductionParameters}
\title{Print AFT productivities}
\usage{
hl_printAgentProductionParameters(simp, filenameprefix = "AftProduction_",
  filenamepostfix = "_multi_medium", aftwisefolder = TRUE,
  heading = "subsubsection", filenameprefix_aftparams = "AftParams_",
  filenamepostfix_aftparams = "")
}
\arguments{
\item{simp}{}

\item{filenameprefix}{}

\item{filenamepostfix}{}
}
\value{
LaTeX subsections and tables
}
\description{
Print AFT productivities
}
\author{
Sascha Holzhauer
}

