% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_params.R
\name{hl_printCompetitionFunctions}
\alias{hl_printCompetitionFunctions}
\title{Read and plot competition functions}
\usage{
hl_printCompetitionFunctions(simp, srcfilename = NULL,
  srcfilepath = dirname(paste(simp$dirs$output$data, simp$sim$folder,
  hl_getRunParameter(simp, "Competition_xml"), sep = "/")), xrange = c(-3, 3),
  yrange = c(-1, 1), filename = "competitionFunctions",
  runidcolumnname = "run", returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{srcfilename}{competition XML file. Will be generated from \code{simp} if missing.}

\item{srcfilepath}{obtained from Links.csv by default. Assign \code{NULL} to apply
\code{simp$dirs$param$getparamdir}.}

\item{xrange}{Vector of two. The x range to plot}

\item{yrange}{Vector of two. The y range to plot}

\item{filename}{filename for figure}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot
}
\description{
Read and plot competition functions
}
\author{
Sascha Holzhauer
}

