% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_printCompetitionFunctions}
\alias{hl_printCompetitionFunctions}
\title{Read and plot compotition functions}
\usage{
hl_printCompetitionFunctions(simp, filename = "Competition_linear",
  xrange = c(-3, 3))
}
\arguments{
\item{simp}{}

\item{filename}{}

\item{xrange}{Vector of two. The x range to plot}
}
\value{
plot
}
\description{
Read and plot compotition functions
}
\author{
Sascha Holzhauer
}

