% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_regions_aftcomposition}
\alias{hl_regions_aftcomposition}
\title{Regional AFT composition}
\usage{
hl_regions_aftcomposition(simp, dataname = "csv_cell_aggregated",
  facet_ncol = 5)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{facet_ncol}{}
}
\value{
plot
}
\description{
Regional AFT composition
}
\author{
Sascha Holzhauer
}

