% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_regions.R
\name{hl_regions_demandandsupply}
\alias{hl_regions_demandandsupply}
\title{Regional demand and supply figure}
\usage{
hl_regions_demandandsupply(simp, runid = simp$sim$runids[1],
  dataname = "csv_cell_aggregated", facet_ncol = 5,
  title = paste("Regional Demand & Supply",
  simp$sim$rundesc[as.character(runid)]), returnplot = FALSE)
}
\arguments{
\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot
}
\description{
Regional demand and supply figure
}
\author{
Sascha Holzhauer
}
