% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_serviceproduction}
\alias{hl_serviceproduction}
\title{Load, aggregate and visualise production per service and AFT}
\usage{
hl_serviceproduction(simp, dataname = "csv_cell_aggregated", facet_ncol = 2,
  normaliseByAftNumber = TRUE, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
timeline plot
}
\description{
Load, aggregate and visualise production per service and AFT
}
\author{
Sascha Holzhauer
}

