% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_data_tables.R
\name{hl_table_demand}
\alias{hl_table_demand}
\title{Output aggregated demand as table with services as columns}
\usage{
hl_table_demand(simp, dataname = "csv_aggregateServiceDemand", latex = TRUE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{latex}{}
}
\value{
data.frame or printed LaTeX table
}
\description{
Output aggregated demand as table with services as columns
}
\author{
Sascha Holzhauer
}

