% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_data_tables.R
\name{hl_table_supply}
\alias{hl_table_supply}
\title{Output aggregated supply as table with services as columns}
\usage{
hl_table_supply(simp, celldataname = "csv_cell_aggregated",
  sourcedataname = "dataAggregateSupplyDemand",
  dataname = "csv_aggregated_supply", latex = TRUE)
}
\arguments{
\item{simp}{}

\item{celldataname}{name for retrieving supply data (celldata)}

\item{sourcedataname}{name for storing supply data}

\item{dataname}{name for retrieving supply data (RData)}

\item{latex}{}
}
\value{
data.frame or printed LaTeX table
}
\description{
Output aggregated supply as table with services as columns
}
\author{
Sascha Holzhauer
}

