% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_takeovers}
\alias{hl_takeovers}
\title{Transition plot of AFT take overs due to giving in.}
\usage{
hl_takeovers(simp, runid = simp$sim$runids[1],
  dataname = "csv_cell_aggregated", starttick = simp$sim$starttick,
  tickinterval = 5, endtick = simp$sim$endtick,
  datanametakeovers = "csv_aggregateTakeOver", transitionthreshold = 5,
  aftnames = simp$mdata$aftNames[-1])
}
\arguments{
\item{simp}{\itemize{
    \item \code{\link{input_tools_load}}
    \item \code{\link{output_visualise_takeovers}}
}}

\item{runid}{}

\item{dataname}{data name of aggregated cell data - used to derive total AFT numbers.}

\item{starttick}{first tick to consider}

\item{tickinterval}{interval between transition columns. Note: tickinterval is based on starttick which defaults to 0!}

\item{endtick}{last tick to consider}

\item{datanametakeovers}{data name for take over data (usually from
\href{https://www.wiki.ed.ac.uk/display/CRAFTY/Output#Output-TakeoverCellOutputter}{CRAFTY outputter TakeoverCellOutputter}).
            and imported via \code{\link{input_csv_data}}).}

\item{transitionthreshold}{minimum number of transitions to show an arrow}

\item{aftnames}{AFT names (defaults to simp$mdata$aftNames without the first entry which is usually
'Unmanaged' and not considered in stored take over data)}
}
\value{
transition plot
}
\description{
Transition plot of AFT take overs due to giving in.
}
\author{
Sascha Holzhauer
}
\seealso{
Other takeovers: \code{\link{convert_aggregate_takeovers}};
  \code{\link{hl_takeovers_all}};
  \code{\link{output_visualise_aftFluctuations}};
  \code{\link{output_visualise_takeovers}}
}

