% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_takeovers}
\alias{hl_takeovers}
\title{Transition plot of AFT take overs due to giving in}
\usage{
hl_takeovers(simp, runid = simp$sim$runids[1],
  dataname = "csv_cell_aggregated", starttick = simp$sim$starttick,
  tickinterval = 5, endtick = simp$sim$endtick,
  datanametakeovers = "csv_aggregateTakeOver", transitionthreshold = 5,
  aftnames = simp$mdata$aftNames[-1])
}
\arguments{
\item{simp}{}

\item{runid}{}

\item{dataname}{}

\item{starttick}{}

\item{tickinterval}{}

\item{endtick}{}

\item{datanametakeovers}{}

\item{transitionthreshold}{}

\item{aftnames}{AFT names (defaults to simp$mdata$aftNames without the first entry which is usually
'Unmanaged' and not considered in stored take over data)}
}
\value{
transition plot
}
\description{
Transition plot of AFT take overs due to giving in
}
\author{
Sascha Holzhauer
}

