% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_takeovers}
\alias{hl_takeovers}
\title{Transition plot of AFT take overs due to giving in.}
\usage{
hl_takeovers(simp, runid = simp$sim$runids[1],
  dataname = "csv_cell_aggregated", starttick = simp$sim$starttick,
  tickinterval = 5, endtick = simp$sim$endtick,
  datanametakeovers = "csv_aggregateTakeOver", transitionthreshold = 5,
  aftnames = simp$mdata$aftNames, aftaggregation = NULL,
  grouping = c("Tick", "Scenario", "Runid", "Region"), aftorder = NULL,
  type_of_arrow = "gradient2sided", returnplot = FALSE)
}
\arguments{
\item{simp}{\itemize{
    \item \code{\link[craftyr]{input_tools_load}}
    \item \code{\link[craftyr]{output_visualise_takeovers}}
}}

\item{runid}{}

\item{dataname}{data name of aggregated cell data - used to derive total AFT numbers.}

\item{starttick}{first tick to consider}

\item{tickinterval}{interval between transition columns. Note: tickinterval is based on starttick which defaults to 0!}

\item{endtick}{last tick to consider}

\item{datanametakeovers}{data name for take over data (usually from 
\href{https://www.wiki.ed.ac.uk/display/CRAFTY/Output#Output-TakeoverCellOutputter}{CRAFTY outputter TakeoverCellOutputter}).
            and imported via \code{\link{input_csv_data}}).}

\item{transitionthreshold}{minimum number of transitions to show an arrow}

\item{aftnames}{the AFTs to show. E.g., use \code{simp$mdata$aftNames[-1]} to omit the first AFT, 
which is usually 'Unmanaged' (and may not be considered in stored take over data)}

\item{aftaggregation}{named vector with names of existing AFT names and values IDs of new groups. Also consider setting
\code{simp$colours$aftgroups} to reflect group colors). To exclude AFT from being displayed, assign NA.}

\item{grouping}{vector of names of columns that describe the dataset but to not contain counts. Column \code{AFT} does not need to be considered here.}

\item{aftorder}{specifies order of buckets in columns from top to bottom. Alphabetical if NULL.}

\item{type_of_arrow}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
transition plot
}
\description{
Transition plot of AFT take overs due to giving in.
}
\author{
Sascha Holzhauer
}
\seealso{
Other takeovers: \code{\link{convert_aggregate_takeovers}},
  \code{\link{hl_takeovers_all}},
  \code{\link{output_visualise_aftFluctuations}},
  \code{\link{output_visualise_takeovers}}
}

