% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_takeovers_all}
\alias{hl_takeovers_all}
\title{Transition plot of AFT take overs (both due to giving in and giving up)}
\usage{
hl_takeovers_all(simp, runid = simp$sim$runids[1],
  dataname = "csv_cell_aggregated", starttick = 2010, tickinterval = 5,
  endtick = 2040, datanametakeovers = "csv_aggregateTakeOver")
}
\arguments{
\item{simp}{}

\item{runid}{}

\item{dataname}{}

\item{starttick}{}

\item{tickinterval}{}

\item{endtick}{}

\item{datanametakeovers}{}
}
\value{
transition plot
}
\description{
Transition plot of AFT take overs (both due to giving in and giving up)
}
\author{
Sascha Holzhauer
}

