% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_aggregate_visualise.R
\name{hl_takeovers_all}
\alias{hl_takeovers_all}
\title{Transition plot of AFT take overs (both due to giving in and giving up).}
\usage{
hl_takeovers_all(simp, runid = simp$sim$runids[1],
  landusedataname = "csv_LandUseIndex_rbinded",
  starttick = simp$sim$starttick, tickinterval = 10,
  endtick = simp$sim$endtick, datanametakeovers = "csv_aggregateTakeOver",
  dataname = "csv_cell_aggregated", grouping = c("Scenario", "Runid",
  "Region"), type_of_arrow = c("gradient2sided", "grid", "simple",
  "gradient"), returnplot = FALSE)
}
\arguments{
\item{simp}{\itemize{
    \item \code{simp$mdata$aftNames}
 \item \code{simp$dirs$output$data}
}}

\item{runid}{}

\item{landusedataname}{data name for land use indices}

\item{starttick}{first tick to consider}

\item{tickinterval}{interval between transition columns. Note: tickinterval is based on starttick which defaults to 0!}

\item{endtick}{last tick to consider}

\item{datanametakeovers}{data name for giving in take overs}

\item{dataname}{data name of aggregated cell data - used to derive total AFT numbers.}

\item{grouping}{passed to \code{\link{convert_aggregate_takeovers}}}

\item{type_of_arrow}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
transition plot
}
\description{
In case you want only take overs due to giving in, considers \code{\link{hl_takeovers}} with data from 
\href{https://www.wiki.ed.ac.uk/display/CRAFTY/Output#Output-TakeoverCellOutputter}{CRAFTY outputter TakeoverCellOutputter}.
}
\details{
According CRAFTY outputter is \href{https://www.wiki.ed.ac.uk/display/CRAFTY/Output#Output-CellTable}{CellTable}.  Applies 
\code{\link{convert_aggregate_takeovers}} to derive take overs from cell data if not yet present.
}
\author{
Sascha Holzhauer
}
\seealso{
Other takeovers: \code{\link{convert_aggregate_takeovers}},
  \code{\link{hl_takeovers}},
  \code{\link{output_visualise_aftFluctuations}},
  \code{\link{output_visualise_takeovers}}
}

