% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hl_volatility}
\alias{hl_volatility}
\title{Read cell volatility data and visualise CellVolatility and NumVolatileCells as lines}
\usage{
hl_volatility(simp, dataname = "csv_aggregated_cellvolatility",
  datatype = "AggregateCellVolatility", datacolumns = c("CellVolatility",
  "NumVolatileCells"), percent = NULL)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{datatype}{}

\item{datacolumns}{}
}
\value{
plot (and rData)
}
\description{
Read cell volatility data and visualise CellVolatility and NumVolatileCells as lines
}
\author{
Sascha Holzhauer
}

