% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_csv_data}
\alias{input_csv_data}
\title{Reads output data from CSV files
Reads output data for the specified datatype amd dataname from CSV data for potentially multiple runs,
regions, and multiple ticks.}
\usage{
input_csv_data(simp, datatype = NULL, dataname = "Cell", colums = NULL,
  pertick = FALSE, extension = "csv", starttick = 0,
  endtick = simp$tech$maxtick, tickinterval = 1, attachfileinfo = TRUE,
  splitfileinfo = FALSE, bindrows = FALSE)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{datatype}{datatype (e.g. "Capital")}

\item{dataname}{dataname (e.g. "Cap1")}

\item{colums}{Vector of colum names. If given restricts returned colums to the given headers (plus X and Y coordinates)}

\item{pertick}{If TRUE the filename will be complemented by all available ticks}

\item{attachfileinfo}{If TRUE, further information about the file is attached to the returned list}

\item{splitfileinfo}{If TRUE, use split upon fileinfo data (not tested)}

\item{bindrows,}{If TRUE, rbind all data into one data.frame}
}
\value{
List of data.frames containing requested data
}
\description{
Reads output data from CSV files
Reads output data for the specified datatype amd dataname from CSV data for potentially multiple runs,
regions, and multiple ticks.
}
\author{
Sascha Holzhauer
}

