% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_csv_param_agents}
\alias{input_csv_param_agents}
\title{Read agent specifc parameters}
\usage{
input_csv_param_agents(simp, aft, filenameprefix = "AftParams_",
  filenamepostfix = "")
}
\arguments{
\item{simp}{}

\item{aft}{}

\item{filenameprefix}{}

\item{filenamepostfix}{}
}
\value{
agent param data
}
\description{
Read agent specifc parameters
}
\author{
Sascha Holzhauer
}

