% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_csv_param_capitals_cellnumbers}
\alias{input_csv_param_capitals_cellnumbers}
\title{Determines number of cells from CSV data for potentially multiple regions}
\usage{
input_csv_param_capitals_cellnumbers(simp, regionpartfromend = 2,
  regionpartdevider = "_")
}
\arguments{
\item{simp}{SIMulation Properties}

\item{regionpartfromend}{index of region part counted from end of filename (used to extract region from filename)}

\item{regionpartdevider}{character that separated the region filename parts (used to extract region from filename)}
}
\value{
data.frame with columns Region and Cells
}
\description{
Determines number of cells from CSV data for potentially multiple regions
}
\author{
Sascha Holzhauer
}

