% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_csv_param_capitals_cellnumbers}
\alias{input_csv_param_capitals_cellnumbers}
\title{Reads capital levels for the specified capitals from CSV data for potentially multiple regions}
\usage{
input_csv_param_capitals_cellnumbers(simp, capitals = simp$mdata$capitals,
  regionpartfromend = 2, regionpartdevider = "_")
}
\arguments{
\item{simp}{SIMulation Properties}

\item{capitals}{vector of strings of requested capital levels}

\item{regionpartfromend}{index of region part counted from end of filename}

\item{regionpartdevider}{character that separated the region filename parts}
}
\value{
data.frame with column Region and Cells
}
\description{
Reads capital levels for the specified capitals from CSV data for potentially multiple regions
}
\author{
Sascha Holzhauer
}

