% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_csv_param.R
\name{input_csv_param_demand}
\alias{input_csv_param_demand}
\title{Reads demand values from CSV data for potentially multiple regions}
\usage{
input_csv_param_demand(simp)
}
\arguments{
\item{simp}{SIMulation Properties. Considered elements:\itemize{
\item{\code{simp$sim$filepartorder}}
\item{\code{simp$sim$scenario}}
\item{\code{simp$sim$regionalisation}}
\item{\code{simp$sim$runids}}
\item{\code{simp$sim$regions}}
\item{\code{simp$dirs$param$getparamdir}}
\item{\code{simp$sim$filepartorder_demands}}}}
}
\value{
list of data.frames containing demand values, and the filename of their origin
}
\description{
Reads demand values from CSV data for potentially multiple regions
}
\author{
Sascha Holzhauer
}
