% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_csv_param.R
\name{input_csv_param_productivities}
\alias{input_csv_param_productivities}
\title{Reads productivities for each capital and each service for the given aft}
\usage{
input_csv_param_productivities(simp, aft, filenameprefix = NULL,
  filenamepostfix = NULL, filenameprefix_aftparams = "AftParams_",
  filenamepostfix_aftparams = "", aftwisefolder = TRUE,
  servicesasrownames = FALSE)
}
\arguments{
\item{simp}{SIMulation Properties}
}
\value{
data.frame containing productivities
}
\description{
Reads productivities for each capital and each service for the given aft
}
\author{
Sascha Holzhauer
}
