% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_csv_param.R
\name{input_csv_param_runs}
\alias{input_csv_param_runs}
\title{Read Runs data (Runs.csv)}
\usage{
input_csv_param_runs(simp, paramid = FALSE)
}
\arguments{
\item{simp}{}
}
\value{
run data
}
\description{
Read Runs data (Runs.csv)
}
\author{
Sascha Holzhauer
}

