% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_various.R
\name{input_marginalutilities}
\alias{input_marginalutilities}
\title{Reads marginal utilties and prepares the data for plotting. Stores data and returns.}
\usage{
input_marginalutilities(simp, filename = paste(simp$dirs$output$rdata,
  "MarginalUtilities.csv", sep = "/"))
}
\arguments{
\item{simp}{}

\item{filename}{}
}
\value{
data.frame
}
\description{
Reads marginal utilties and prepares the data for plotting. Stores data and returns.
}
\author{
Sascha Holzhauer
}

