% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_raster.R
\name{input_raster_output}
\alias{input_raster_output}
\title{Get list (scenarios/regions/runids) of lists (ticks) of raster data from files whose source is taken from simp.}
\usage{
input_raster_output(simp, datatype = NULL, dataname, starttick = if
  (!is.null(simp$sim$starttick)) simp$sim$starttick else simp$tech$mintick,
  endtick = if (!is.null(simp$sim$endtick)) simp$sim$endtick else
  simp$tech$maxtick, tickinterval = 1, encapsefileinfo = TRUE)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{datatype}{category of raster data (e.g. "Capital")}

\item{dataname}{name of data of category (e.g. "Cap1")}

\item{starttick}{first tick to consider}

\item{endtick}{last tick to consider}

\item{tickinterval}{Note: tickinterval is based on starttick which defaults to 0!}
}
\value{
list of lists of raster data
}
\description{
Get list (scenarios/regions/runids) of lists (ticks) of raster data from files whose source is taken from simp.
}
\author{
Sascha Holzhauer
}

