% Generated by roxygen2 (4.0.1): do not edit by hand
\name{input_raster_output}
\alias{input_raster_output}
\title{Get list (regions) of lists (ticks) of raster data from files whose source is taken from simp.}
\usage{
input_raster_output(simp, datatype, dataname, starttick = NULL,
  endtick = NULL, tickinterval = 1)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{datatype}{category of raster data (e.g. "Capital")}

\item{dataname}{name of data of category (e.g. "Cap1")}

\item{starttick}{}

\item{endtick}{}

\item{tickinterval}{}
}
\value{
list of lists of raster data
}
\description{
Get list (regions) of lists (ticks) of raster data from files whose source is taken from simp.
}
\author{
Sascha Holzhauer
}

