% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_shapefile.R
\name{input_shapes_countries}
\alias{input_shapes_countries}
\title{Generates a ggplot2 addon to draw (country) shapes}
\usage{
input_shapes_countries(simp, filename_shapes = system.file("extdata",
  "various/europeshapes/Europe_Countries_Final_Projection.shp", package =
  "craftyr"), countries2show = NULL, countrycodedatacolumn = "ISO_CC")
}
\arguments{
\item{simp}{}

\item{filename_shapes}{(Default: shapes from Calum > less accurate than shapes from raster but much smaller)}
}
\value{
ggplot2 addon
}
\description{
Generates a ggplot2 addon to draw (country) shapes
}
\seealso{
http://spatial.ly/2013/12/introduction-spatial-data-ggplot2/
}
\author{
Sascha Holzhauer
}
