% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_tools_getAvailableTicks}
\alias{input_tools_getAvailableTicks}
\title{Determine a vector of ticks for the files in the given directory that match the given pattern.
If pattern is not specified it is assumed that all requested files in the directory are of the form
*<scenario>-<runid>-<region>-<datatype>-<dataname>-<tick>.<extension>.
If extension is not given, any extension is considered. If datatype and dataname are
not given, the according parts of the filename is assumed not to exist (also the associated '-').}
\usage{
input_tools_getAvailableTicks(simp, dir, pattern = NULL, datatype = NULL,
  dataname = NULL, extension = NULL, starttick = 0,
  endtick = simp$tech$maxtick, tickinterval = 1)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{dir}{directory to search in}

\item{pattern}{pattern for matching considered files excluding the tick and extension. If NULL, the default is used (\\.*<scenario>-<runid>-<region>-<datatype>-<dataname>-).}

\item{datatype}{(e.g. "Capital")}

\item{dataname}{(e.g. "Cap1")}

\item{extension}{used to restrict filenames by extension (without "."!)}

\item{starttick}{start tick of requested range of ticks}

\item{endtick}{end tick of requested range of ticks}

\item{tickinterval}{interval of requested range of ticks}
}
\value{
vector of ticks
}
\description{
Determine a vector of ticks for the files in the given directory that match the given pattern.
If pattern is not specified it is assumed that all requested files in the directory are of the form
*<scenario>-<runid>-<region>-<datatype>-<dataname>-<tick>.<extension>.
If extension is not given, any extension is considered. If datatype and dataname are
not given, the according parts of the filename is assumed not to exist (also the associated '-').
}
\author{
Sascha Holzhauer
}

