% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_getAvailableTicks}
\alias{input_tools_getAvailableTicks}
\title{Determine a vector of ticks for the files in the given directory that match the given pattern}
\usage{
input_tools_getAvailableTicks(simp, dir, pattern = NULL, datatype = NULL,
  dataname = NULL, extension = NULL, starttick = if
  (!is.null(simp$sim$starttick)) simp$sim$starttick else simp$tech$mintick,
  endtick = if (!is.null(simp$sim$endtick)) simp$sim$endtick else
  simp$tech$maxtick, tickinterval = 1)
}
\arguments{
\item{simp}{SIMulation Properties
\itemise{
    \item{simp$dirs$output$simulation}
    \item{simp$sim$world}
    \item{simp$sim$regionalisation -can be \code{NULL}}
    \item{simp$sim$scenario}
    \item{simp$sim$runids}
    \item{simp$sim$regions}
    \item{simp$sim$hasregiondir}
    \item{simp$sim$regions}}}

\item{dir}{directory to search in}

\item{pattern}{pattern for matching considered files including the tick (as TICK) and excluding the extension. 
If NULL, the default is used (\\.*<scenario>-<runid>-<region>-<datatype>-<dataname>-TICK).}

\item{datatype}{(e.g. "Capital")}

\item{dataname}{(e.g. "Cap1")}

\item{extension}{used to restrict filenames by extension (without "."!)}

\item{starttick}{start tick of requested range of ticks}

\item{endtick}{end tick of requested range of ticks}

\item{tickinterval}{interval of requested range of ticks}
}
\value{
vector of ticks
}
\description{
If pattern is not specified it is assumed that all requested files in the directory are of the form 
*<scenario>-<runid>-<region>-<datatype>-<dataname>-<tick>.<extension>.
If extension is not given, any extension is considered. If datatype and dataname are 
not given, the according parts of the filename is assumed not to exist (also the associated '-').
}
\author{
Sascha Holzhauer
}
