% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_getModelInputDir}
\alias{input_tools_getModelInputDir}
\title{Determines the model input folder for the given datadir}
\usage{
input_tools_getModelInputDir(simp, datatype = NULL)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{datatype}{one of c("capitals", "demand", "agentparams", "productivities", "competition", "runs")}
}
\value{
model output directory as string
}
\description{
Determines the model input folder for the given datadir
}
\author{
Sascha Holzhauer
}

