% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_getModelInputFilenames}
\alias{input_tools_getModelInputFilenames}
\title{Determines the model input filenames for the given simp and datatype, dataname, ticks settings}
\usage{
input_tools_getModelInputFilenames(simp,
  folders = simp$dirs$param$getparamdir(simp, datatype), datatype = NULL,
  dataname = NULL, extension = NULL, pertick = FALSE, starttick = NULL,
  endtick = NULL, tickinterval = NULL, returnfileinfo = TRUE)
}
\arguments{
\item{simp}{}

\item{folders}{}

\item{datatype}{}

\item{dataname}{}

\item{extension}{}

\item{pertick}{if TRUE the filename will be complemented by all available ticks
(using \code{\link{input_tools_getAvailableTicks}})}

\item{starttick}{only required when \code{pertick == TRUE}}

\item{endtick}{only required when \code{pertick == TRUE}}

\item{tickinterval}{only required when \code{pertick == TRUE}}

\item{returnfileinfo}{}
}
\value{
list of vector of filenames (list elements rerpesent files of one folder)
}
\description{
Determines the model input filenames for the given simp and datatype, dataname, ticks settings
}
\author{
Sascha Holzhauer
}

