% Generated by roxygen2 (4.0.1): do not edit by hand
\name{input_tools_getModelOutputDir}
\alias{input_tools_getModelOutputDir}
\title{Determines the model output folder for the given simp}
\usage{
input_tools_getModelOutputDir(simp)
}
\arguments{
\item{simp}{SIMulation Properties}
}
\value{
String of output folder
}
\description{
Determines the model output folder for the given simp
}
\author{
Sascha Holzhauer
}

