% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_tools_getModelOutputFilenames}
\alias{input_tools_getModelOutputFilenames}
\title{Determines the model output filenames for the given simp and datatype, dataname, ticks settings.
Able to read multiple scenarios/regions/runids.
Attaches infos (region, runid, scenario).}
\usage{
input_tools_getModelOutputFilenames(simp,
  folders = input_tools_getModelOutputDir(simp), datatype = NULL,
  dataname = NULL, extension = NULL, returnfileinfo = TRUE,
  pertick = FALSE, starttick = 0, endtick = simp$tech$maxtick,
  tickinterval = 1)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{folders}{}

\item{datatype}{(e.g. "Capital")}

\item{dataname}{(e.g. "Cap1")}

\item{extension}{used to restrict filenames by extension (without "."!)}

\item{pertick}{if TRUE the filename will be complemented by all available ticks}

\item{starttick}{start tick of requested range of ticks}

\item{endtick}{end tick of requested range of ticks}

\item{tickinterval}{interval of requested range of ticks}
}
\value{
list of vector of filenames (list elements rerpesent files of one folder)
}
\description{
Determines the model output filenames for the given simp and datatype, dataname, ticks settings.
Able to read multiple scenarios/regions/runids.
Attaches infos (region, runid, scenario).
}
\author{
Sascha Holzhauer
}

