% Generated by roxygen2 (4.0.2): do not edit by hand
\name{input_tools_load}
\alias{input_tools_load}
\title{Wrapper for load}
\usage{
input_tools_load(simp, objectName, ...)
}
\arguments{
\item{simp}{}

\item{objectName}{}
}
\description{
Wrapper for load
}
\author{
Sascha Holzhauer
}

