% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{input_tools_load}
\alias{input_tools_load}
\title{Wrapper for load}
\usage{
input_tools_load(simp, objectName, ...)
}
\arguments{
\item{simp}{\itemize{
    \item \code{simp$sim$id}
 \item \code{simp$dirs$output$rdata}
    \item \code{simp$sim$version} (if \code{is.null(simp$sim$id)})
}}

\item{objectName}{}
}
\description{
Wrapper for load
}
\author{
Sascha Holzhauer
}

