% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_xml_param.R
\name{input_xml_param_competition}
\alias{input_xml_param_competition}
\title{Reads competition functions for each service.}
\usage{
input_xml_param_competition(simp, srcfilepath = NULL,
  srcfilename = "Competition_linear", returntext = FALSE)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{srcfilepath}{path for competition XML file. Applies \code{simp$dirs$param$getparamdir}
to obtain path if NULL.}

\item{srcfilename}{Filename of competition XML file without extention.}
}
\value{
data.frame containing productivities
}
\description{
Reads competition functions for each service.
}
\author{
Sascha Holzhauer
}
