% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_configuration.R
\name{make_demand_static}
\alias{make_demand_static}
\title{Copy first line of CSV file and assign endtick.}
\usage{
make_demand_static(sip, indir = simp$dirs$param$getparamdir(simp, "demand"),
  outdir = indir, endtick = simp$sim$endtick, tickcolumnname = "Year",
  filepattern = "*.csv")
}
\arguments{
\item{sip}{}

\item{indir}{}

\item{outdir}{}

\item{endtick}{}
}
\value{
altered CSV files
}
\description{
Copy first line of CSV file and assign endtick.
}
\author{
Sascha Holzhauer
}
