% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_takeovers.R
\name{output_visualise_aftFluctuations}
\alias{output_visualise_aftFluctuations}
\title{Visualise AFT fluctuations}
\usage{
output_visualise_aftFluctuations(simp, data, starttick = if
  (!is.null(simp$sim$starttick)) simp$sim$starttick else simp$tech$mintick,
  endtick = if (!is.null(simp$sim$endtick)) simp$sim$endtick else
  simp$tech$maxtick, tickinterval = 1, title = "AFT Fluctuations",
  filename = title, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{data}{}

\item{starttick}{}

\item{endtick}{}

\item{tickinterval}{}

\item{title}{}

\item{filename}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot
}
\description{
Visualise AFT fluctuations
}
\author{
Sascha Holzhauer
}
\seealso{
Other takeovers: \code{\link{convert_aggregate_takeovers}},
  \code{\link{hl_takeovers_all}},
  \code{\link{hl_takeovers}},
  \code{\link{output_visualise_takeovers}}
}

