% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_initfigure.R
\name{output_visualise_initFigure}
\alias{output_visualise_initFigure}
\title{Init figure output}
\usage{
output_visualise_initFigure(simp, outdir = simp$dirs$output$figures, filename,
  ensurePath = TRUE)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{outdir}{(default: simp$dirs$output$figures)}

\item{filename}{}

\item{ensurePath}{switch}
}
\value{
plot
}
\description{
Init figure output
}
\author{
Sascha Holzhauer
}

