% Generated by roxygen2 (4.0.2): do not edit by hand
\name{output_visualise_takeovers}
\alias{output_visualise_takeovers}
\title{Visualise takeovers of land uses as transition plot}
\usage{
output_visualise_takeovers(simp, data, startpopulation = NULL,
  starttick = 0, endtick = simp$tech$maxtick, tickinterval = 1,
  type_of_arrow = "grid", transitionthreshold = 0)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{data}{takeover data with column "Tick" and one for each AFT}

\item{startpopulation}{needs to contain a column 'Number' which contains agent numbers of the start tick}

\item{starttick}{first tick to consider}

\item{endtick}{last tick to consider}

\item{tickinterval}{Note: tickinterval is based on starttick which defaults to 0!}

\item{type_of_arrow}{}

\item{transitionThreshold}{}
}
\value{
plots figure
}
\description{
Visualise takeovers of land uses as transition plot
}
\author{
Sascha Holzhauer
}

