% Generated by roxygen2 (4.0.2): do not edit by hand
\name{param_getDefaultSimp}
\alias{param_getDefaultSimp}
\title{Get the default SIMP parameter list}
\usage{
param_getDefaultSimp()
}
\value{
list of default SIMP parameters
}
\description{
Get the default SIMP parameter list
}
\author{
Sascha Holzhauer
}

