% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_raster_getMainFilename}
\alias{read_raster_getMainFilename}
\title{Concenate the filename from basedir, scenario, world, run_id, year as well as dataType and dataName.
Region, dataType, and dataName are only considered if given as parameters.}
\usage{
read_raster_getMainFilename(basedir, scenario, world, run_id, region = NULL,
  dataType = NULL, dataName = NULL, year, useRegionDir = FALSE)
}
\arguments{
\item{scenario}{}

\item{world}{}

\item{run_id}{}

\item{year}{}

\item{basedir}{}

\item{dataType}{(e.g. "Capital")}

\item{dataName}{(e.g. "Cap1")}

\item{useRegionDir}{(output data may or may not be stored in region specific folders)}
}
\value{
String of filename
}
\description{
Concenate the filename from basedir, scenario, world, run_id, year as well as dataType and dataName.
Region, dataType, and dataName are only considered if given as parameters.
}
\author{
Calum Brown

Sascha Holzhauer
}

