% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_raster_getMainFilename}
\alias{read_raster_getMainFilename}
\title{Concenate the filename from basedir, scenario, world, run_id, year as well as datatype and dataname.
Region, datatype, and dataname are only considered if given as parameters.}
\usage{
read_raster_getMainFilename(basedir, scenario, world, run_id, region = NULL,
  dataType = NULL, dataName = NULL, year, useregiondir = FALSE)
}
\arguments{
\item{basedir}{}

\item{scenario}{}

\item{world}{}

\item{run_id}{}

\item{year}{}

\item{useregiondir}{(output data may or may not be stored in region specific folders)}

\item{datatype}{(e.g. "Capital")}

\item{dataname}{(e.g. "Cap1")}
}
\value{
String of filename
}
\description{
Concenate the filename from basedir, scenario, world, run_id, year as well as datatype and dataname.
Region, datatype, and dataname are only considered if given as parameters.
}
\author{
Calum Brown

Sascha Holzhauer
}

