% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readingRasterData_collateRuns}
\alias{readingRasterData_collateRuns}
\title{Read raster}
\usage{
readingRasterData_collateRuns(scenario = lix$scenario, world = pix$world,
  basedir = pix$path_names["dataSource"][[1]], region = NULL,
  num_runs = NULL, dataType, dataName, aggregate = FALSE,
  runIDrange = NULL, useRegionDir = FALSE, func = NULL,
  startYear = NULL, endYear = NULL, everyYear = NULL)
}
\arguments{
\item{scenario}{}

\item{world}{}

\item{basedir}{}

\item{region}{}

\item{num_runs}{the number of runs to consider}

\item{dataType}{}

\item{dataName}{}

\item{runIDrange}{}

\item{useRegionDir}{(output data may or may not be stored in region specific folders)}

\item{func}{}

\item{endYear}{}

\item{everyYear}{}

\item{startyear}{}
}
\value{
data.frame
}
\description{
Read raster
}
\author{
Sascha Holzhauer
}

