% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_raster.R
\name{readingRasterData_collateRunsAndMeasures}
\alias{readingRasterData_collateRunsAndMeasures}
\title{Collate runs and measure in data.frame}
\usage{
readingRasterData_collateRunsAndMeasures(scenario = lix$scenario,
  world = pix$world, basedir = pix$path_names["dataSource"][[1]],
  region = NULL, num_runs = NULL, dataTypes, dataNames, aggregate = TRUE,
  runIDrange = NULL, useRegionDir = FALSE, funcs = NULL,
  startYear = lix$startTick, endYear = lix$endTick,
  everyYear = lix$everyTick)
}
\arguments{
\item{dataTypes}{vector of dataTypes}

\item{dataNames}{according vector of dataName with same length as dataType}

\item{useRegionDir}{(output data may or may not be stored in region specific folders)}
}
\value{
data.frame
}
\description{
Collate runs and measure in data.frame
}
\author{
Sascha Holzhauer
}
