% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readRasterDataYears}
\alias{readRasterDataYears}
\title{Read a number of raster files and combine them in a list of data.}
\usage{
readRasterDataYears(scenario = lix$scenario, world = pix$world,
  basedir = pix$path_names["dataSource"][[1]], run_id = lix$run_id,
  startYear = lix$startTick, endYear = lix$endTick,
  everyYear = lix$everyTick, region = lix$region, dataType = NULL,
  dataName = "Agent-SerialID", useRegionDir = FALSE)
}
\arguments{
\item{scenario}{}

\item{world}{}

\item{basedir}{}

\item{run_id}{}

\item{startYear}{}

\item{endYear}{}

\item{everyYear}{}

\item{region}{}

\item{dataType}{}

\item{dataName}{}

\item{useRegionDir}{(output data may or may not be stored in region specific folders)}
}
\value{
list of data over years
}
\description{
Read a number of raster files and combine them in a list of data.
}
\author{
Sascha Holzhauer
}

