% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_colours.R
\name{settings_colours_getAftColours}
\alias{settings_colours_getAftColours}
\title{Return AFT colours}
\usage{
settings_colours_getAftColours()
}
\value{
coolour pallete
}
\description{
Return AFT colours
}
\author{
Sascha Holzhauer
}

