% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_colours.R
\name{settings_colours_getColors}
\alias{settings_colours_getColors}
\title{Get colour palette of requested size.}
\usage{
settings_colours_getColors(simp, number, dim = 1, set = "Set1")
}
\arguments{
\item{number}{}

\item{dim}{}
}
\value{
colour palette
}
\description{
Applied if colours are not defined in \code{simp$fills} or \code{simp$colours}
}
\author{
Sascha Holzhauer
}

