% Generated by roxygen2 (4.0.2): do not edit by hand
\name{settings_colours_getColors}
\alias{settings_colours_getColors}
\title{Get colour palette of requested dimension}
\usage{
settings_colours_getColors(number, dim = 1)
}
\arguments{
\item{number}{}

\item{dim}{}
}
\value{
colour palette
}
\description{
Get colour palette of requested dimension
}
\author{
Sascha Holzhauer
}

