% Generated by roxygen2 (4.0.2): do not edit by hand
\name{settings_colours_getColorSet}
\alias{settings_colours_getColorSet}
\title{Return requested colour palette}
\usage{
settings_colours_getColorSet(number = 9, set = "Set1")
}
\arguments{
\item{number}{number of required colours}

\item{set}{one of c("Topo","RedGreen","GreenRed") or one for brewer.pal}
}
\value{
colour palette
}
\description{
Return requested colour palette
}
\author{
Sascha Holzhauer
}

