% Generated by roxygen2 (4.0.2): do not edit by hand
\name{settings_colours_getServiceColours}
\alias{settings_colours_getServiceColours}
\title{Return AFT colours}
\usage{
settings_colours_getServiceColours()
}
\value{
coolour pallete
}
\description{
Return AFT colours
}
\author{
Sascha Holzhauer
}

