% Generated by roxygen2 (4.0.2): do not edit by hand
\name{settings_colours_showColours}
\alias{settings_colours_showColours}
\title{Show colour palette}
\usage{
settings_colours_showColours(bg = "gray", cex = 0.75, srt = 30)
}
\arguments{
\item{bg}{}

\item{cex}{}

\item{srt}{}
}
\value{
colour palettes plot
}
\description{
Show colour palette
}
\author{
Sascha Holzhauer
}

