% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visualisation_raster_legendonlytheme}
\alias{visualisation_raster_legendonlytheme}
\title{Only facet elements}
\usage{
visualisation_raster_legendonlytheme(base_size = 11,
  base_family = "Helvetica")
}
\arguments{
\item{base_size}{}

\item{base_family}{}
}
\value{
theme
}
\description{
Only facet elements
}
\author{
Sascha Holzhauer
}

