% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_raster.R
\name{visualisation_raster_nobackground}
\alias{visualisation_raster_nobackground}
\title{white background}
\usage{
visualisation_raster_nobackground(base_size = 11, base_family = "Helvetica")
}
\arguments{
\item{base_size}{}

\item{base_family}{}
}
\value{
theme
}
\description{
white background
}
\author{
Sascha Holzhauer
}
