% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_raster.R
\name{visualisation_raster_printRawPlots_theme_nothing}
\alias{visualisation_raster_printRawPlots_theme_nothing}
\title{GGplot2 theme for visualisation_raster_printRawPlots}
\usage{
visualisation_raster_printRawPlots_theme_nothing(base_size = 12,
  base_family = "Helvetica")
}
\value{
ggplot2 theme
}
\description{
GGplot2 theme for visualisation_raster_printRawPlots
}
\author{
Sascha Holzhauer
}

