% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_actions.R
\name{visualise_actions}
\alias{visualise_actions}
\title{Visualise the performance and effect of Actions}
\usage{
visualise_actions(simp, dataname = "dataActions",
  monitordataname = "dataAggregateConnectivity",
  monitorvars = simp$mdata$aftNames[-1], monitorColours = NULL,
  title = "Institutional Action", filename = NULL, x_column = "Tick",
  y_column = "Score", colour_column = "FR", shape_column = "Agent",
  fill_column = "Action", facet_column = "Region",
  facet_column_levels = NULL, facet_ncol = 1, alpha_column = "Selected",
  alpha = 0.5, size_column = "RestrictionsNumber",
  measure_name = "Connectedness", actionColours = NULL,
  returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{dataname}{}

\item{monitordataname}{}

\item{title}{}

\item{filename}{}

\item{y_column}{}

\item{colour_column}{}

\item{shape_column}{}

\item{fill_column}{}

\item{facet_column}{}

\item{facet_ncol}{}

\item{alpha_column}{}

\item{size_column}{}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
figure
}
\description{
Visualise the performance and effect of Actions
}
\author{
Sascha Holzhauer
}

