% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visualise_cells_printRawPlots}
\alias{visualise_cells_printRawPlots}
\title{Writes a list of raster data as single raw (only the raster data) ggplot2 plot files}
\usage{
visualise_cells_printRawPlots(simp, celldata, idcolumn = "Tick",
  valuecolumn = "LandUseIndex", title = "", filenamepostfix = title,
  factorial = FALSE, ncol = if (!is.data.frame(celldata)) length(celldata)
  else 1, coloursetname = simp$colours$defaultset, ggplotaddon = NULL)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{factorial}{true if raster values are factorial (affects colour palette)}

\item{ncol}{number of columns of facet wrap}

\item{coloursetname}{id for colour set (if factorial) to pass to simp$colours$GenericFun (e.g. "AFT", "Capital", "Service")}

\item{rasterdata}{if a list, elements must be named differently}

\item{datanames}{vector of names of the same length as rasterData}

\item{omitaxisticks}{omit axis ticks if true}
}
\value{
raster visualisation files
}
\description{
Writes a list of raster data as single raw (only the raster data) ggplot2 plot files
}
\examples{
#######################################################################
# crafty demo for raster visualisation
# Last update: 	23.06.2014
# Author: 		Sascha Holzhauer
#######################################################################
source(simp$simpDefinition)  # read simp
data <- input_csv_param_capitals()
raster <- convert_2raster(simp, data)
visualisation_raster_printRawPlots(raster, legendTitle = "Capitals Example", 
		factorial= FALSE, omitAxisTicks = FALSE, ncol = 1, id="None")
}
\author{
Sascha Holzhauer
}

