% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_competition.R
\name{visualise_competition_funcs}
\alias{visualise_competition_funcs}
\title{Plot competition functions for services}
\usage{
visualise_competition_funcs(simp, functions, xrange = c(-3, 3),
  yrange = c(-1, 1), filename = "competitionFunctions",
  returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{functions}{}

\item{xrange}{Vector of two. The x range to plot}

\item{yrange}{Vector of two. The y range to plot}

\item{filename}{filename for figure}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
plot if \code{returnplot == true}
}
\description{
Plot competition functions for services
}
\author{
Sascha Holzhauer
}
