% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visualise_competition_funcs}
\alias{visualise_competition_funcs}
\title{Plot competition functions for services}
\usage{
visualise_competition_funcs(simp, functions, xrange = c(-3, 3))
}
\arguments{
\item{simp}{}

\item{functions}{}

\item{xrange}{Vector of two. The x range to plot}
}
\value{
plot
}
\description{
Plot competition functions for services
}
\author{
Sascha Holzhauer
}

