% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_competition.R
\name{visualise_competition_prealloc}
\alias{visualise_competition_prealloc}
\title{Histogram of competitiveness per AFT: stacked: below/above GU threshold}
\usage{
visualise_competition_prealloc(simp, data,
  facet_ncol = length(simp$mdata$aftNames) - 1,
  filename = paste("PreAllocationCompetition", simp$sim$id, sep = "_"),
  numbins = 20,
  binwidth = diff(range(data$PreAllocCompetitiveness))/numbins,
  title = NULL, ggplotaddons = NULL, setfigdims = TRUE,
  storename = NULL, returnplot = FALSE)
}
\arguments{
\item{simp}{}

\item{facet_ncol}{if \code{NULL} ticks will be as rows, agents as columns. Otherwise, facets will 
distinguished by ticks only (and facet_cols specifies the number of columns)}

\item{filename}{}

\item{numbins}{number of bins to bin data}

\item{title}{plot title}

\item{ggplotaddons}{list of ggplot2 elements}

\item{setfigdims}{if \code{TRUE} \code{simp$fig$height} and \code{simp$fig$width} are set appropriately}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
facet histogram plot
}
\description{
Histogram of competitiveness per AFT: stacked: below/above GU threshold
}
\author{
Sascha Holzhauer
}
