% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visualise_raster_printPlots}
\alias{visualise_raster_printPlots}
\title{Prints a list of raster data as ggplot2 facet plot}
\usage{
visualise_raster_printPlots(simp, inforasterdata, idcolumn = "Tick",
  title = "", filenamepostfix = title, legendtitle = "",
  factorial = FALSE, omitaxisticks = FALSE, ncol = if
  (is.list(inforasterdata)) length(inforasterdata[[1]][, 1]) else
  length(inforasterdata[, 1]), coloursetname = simp$colours$defaultset,
  legenditemnames = NULL, ggplotaddon = NULL,
  theme = visualisation_raster_nobackground)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{inforasterdata}{(list of) data.frames contain info and (if multiple: list of!) raster objects. If a list of data.frames,
elements must be named differently}

\item{idcolumn}{column used to separate and name rasters, refering to column names (set with colnames()) of the data.frame(s).}

\item{title}{name of plot}

\item{filenamepostfix}{appended to the default output filename}

\item{legendtitle}{title for legend of raster values}

\item{factorial}{true if raster values are factorial (affects colour palette)}

\item{omitaxisticks}{omit axis ticks if true}

\item{ncol}{number of columns of facet wrap. Defauls to the number of rasters in the first dataframe}

\item{coloursetname}{id for colour set (if factorial) to pass to simp$colours$GenericFun (e.g. "AFT", "Capital", "Service")}

\item{legenditemnames}{names for legend items}
}
\value{
raster visualisation
}
\description{
Prints a list of raster data as ggplot2 facet plot
}
\examples{
#######################################################################
# crafty demo for raster visualisation (capital input data)
# Last update: 	23.06.2014
# Author: 		Sascha Holzhauer
# TODO adapt from rawPlot
#######################################################################
source(simp$simpDefinition)   # read simp
data <- input_csv_param_capitals()
raster <- convert_2raster(simp, data)
visualisation_raster_printRawPlots(raster, legendTitle = "Capitals Example", 
		factorial= FALSE, omitAxisTicks = FALSE, ncol = 1, id="None")
}
\author{
Sascha Holzhauer
}
\seealso{
output_tools_getDefaultFilename
}

