% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualise_raster.R
\name{visualise_raster_printRawPlots}
\alias{visualise_raster_printRawPlots}
\title{Writes a list of raster data as single raw (only the raster data) ggplot2 plot files}
\usage{
visualise_raster_printRawPlots(simp, rasterdata, datanames = NULL,
  legendtitle = "", factorial = FALSE, omitaxisticks = FALSE, ncol = 1,
  coloursetname = simp$colours$defaultset, legenditemnames = NULL,
  theme = visualisation_raster_legendonlytheme, returnplot = FALSE)
}
\arguments{
\item{simp}{SIMulation Properties}

\item{rasterdata}{if a list, elements must be named differently}

\item{datanames}{vector of names of the same length as rasterData}

\item{legendtitle}{title for legend of raster values}

\item{factorial}{true if raster values are factorial (affects colour palette)}

\item{omitaxisticks}{omit axis ticks if true}

\item{ncol}{number of columns of facet wrap}

\item{coloursetname}{id for colour set (if factorial) to pass to simp$colours$GenericFun (e.g. "AFT", "Capital", "Service")}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
raster visualisation files
}
\description{
There does not seem to be a straigh-forward way to convert a gTree object back to a ggplot2 object...
(http://stackoverflow.com/questions/29583849/r-saving-a-plot-in-an-object)
}
\examples{
#######################################################################
# crafty demo for raster visualisation
# Last update: 	23.06.2014
# Author: 		Sascha Holzhauer
#######################################################################
source(simp$simpDefinition)  # read simp
data <- input_csv_param_capitals()
raster <- convert_2raster(simp, data)
visualisation_raster_printRawPlots(raster, legendTitle = "Capitals Example", 
		factorial= FALSE, omitAxisTicks = FALSE, ncol = 1, id="None")
}
\author{
Sascha Holzhauer
}
